moved functions to src folder, merged similar functions, and edited cleandata_descbr.csv to cleandata_describe.csv
