+++ reports/pdf_malware_detector_report.qmd
@@ -102,7 +102,7 @@ Markdown(cleandata_info_table.to_markdown(index=False))
 #| label: tbl-cleandata-describe
 #| tbl-cap: The summary numerical values in clean data for count, mean, median, quantiles, minimum, maximum, and standard deviation.
 
-cleandata_desc_table = pd.read_csv("../results/tables/cleandata_descr.csv")
+cleandata_desc_table = pd.read_csv("../results/tables/cleandata_describe.csv")
 cleandata_desc_table = cleandata_desc_table.rename(columns={cleandata_desc_table.columns[0]: "Summary Stats"})
 Markdown(cleandata_desc_table.to_markdown(index=False))
 ```
