+++ results/tables/dummy_cv_results.csv
@@ -1,6 +1,6 @@
 ,mean,std
 fit_time,0.002,0.0
-score_time,0.01,0.0
+score_time,0.011,0.0
 test_accuracy,0.554,0.0
 train_accuracy,0.554,0.0
 test_precision,0.554,0.0
