+++ scripts/01-fetch.py
@@ -16,104 +16,14 @@ import glob
 import shutil
 import pandas as pd
 import click
+from src.create_directories import create_directories
+from src.download_and_copy_dataset import download_and_copy_dataset
+from src.export_data_summary import export_data_summary
 
 # Default configuration constants
 DEFAULT_LINK = "dhoogla/cic-evasive-pdfmal2022"
 DEFAULT_DATA_SAVE_FOLDER = "../data/Raw"
 DEFAULT_TABLES_FOLDER = "../results/tables"
-
-
-def create_directories(data_save_folder: str, tables_folder: str) -> None:
-    """
-    Create necessary directories for data storage and results.
-
-    Creates directories if they don't already exist. Uses exist_ok=True to
-    prevent errors if directories already exist.
-
-    Args:
-        data_save_folder (str): Path to the folder where raw data will be stored.
-        tables_folder (str): Path to the folder where CSV results will be stored.
-
-    Returns:
-        None
-    """
-    os.makedirs(data_save_folder, exist_ok=True)
-    os.makedirs(tables_folder, exist_ok=True)
-
-
-def download_and_copy_dataset(link: str, data_save_folder: str) -> tuple[pd.DataFrame, str]:
-    """
-    Download dataset from Kaggle and copy to local storage.
-
-    Downloads the dataset using kagglehub and copies the first parquet file found
-    to the specified data folder. Validates that the copied file is in parquet format.
-
-    Args:
-        link (str): Kaggle dataset link in format "username/dataset-name".
-        data_save_folder (str): Path to the folder where the parquet file will be saved.
-
-    Returns:
-        tuple[pd.DataFrame, str]: A tuple containing:
-            - The loaded DataFrame from the parquet file
-            - The path to the saved parquet file
-
-    Raises:
-        AssertionError: If the saved file is not in parquet format.
-        IndexError: If no parquet files are found in the downloaded dataset.
-    """
-    # Download dataset from Kaggle - returns the local path to downloaded files
-    downloaded_path = kagglehub.dataset_download(link)
-
-    # Find the first parquet file in the downloaded dataset
-    downloaded_parquet_files = glob.glob(os.path.join(downloaded_path, "*.parquet"))
-    if not downloaded_parquet_files:
-        raise FileNotFoundError(f"No parquet files found in {downloaded_path}")
-
-    downloaded_parquet = downloaded_parquet_files[0]
-
-    # Copy the parquet file to our designated data folder
-    data_path = os.path.join(data_save_folder, "PDFMalware2022.parquet")
-    shutil.copy(downloaded_parquet, data_path)
-
-    # Validate that the saved file is indeed in parquet format
-    assert os.path.splitext(data_path)[1] == ".parquet", \
-        f"Expected parquet file, but got {os.path.splitext(data_path)[1]}"
-
-    # Load and return the dataset
-    data = pd.read_parquet(data_path, engine="pyarrow")
-    return data, data_path
-
-
-def export_data_summary(data: pd.DataFrame, tables_folder: str) -> None:
-    """
-    Export data information and statistical summaries to CSV files.
-
-    Generates two CSV files:
-        1. rawdata_info.csv: Column data types and null/non-null counts
-        2. rawdata_describe.csv: Statistical summaries (mean, std, min, max, etc.)
-
-    Args:
-        data (pd.DataFrame): The DataFrame to analyze and export statistics for.
-        tables_folder (str): Path to the folder where CSV files will be saved.
-
-    Returns:
-        None
-    """
-    # Export column-level information (data types and null counts)
-    info_df = pd.DataFrame(
-        {
-            "dtype": data.dtypes,
-            "non_null_count": data.notna().sum(),
-            "null_count": data.isna().sum(),
-        }
-    )
-    info_path = os.path.join(tables_folder, "rawdata_info.csv")
-    info_df.to_csv(info_path)
-
-    # Export descriptive statistics (mean, std, quartiles, etc.)
-    descr_df = data.describe()
-    descr_path = os.path.join(tables_folder, "rawdata_describe.csv")
-    descr_df.to_csv(descr_path)
 
 
 @click.command()
@@ -162,7 +72,7 @@ def main(link: str, data_save_folder: str, tables_folder: str) -> None:
     print(f"[INFO] Data downloaded and saved to {data_path}")
 
     # Step 3: Export data summaries and statistics
-    export_data_summary(data, tables_folder)
+    export_data_summary(data, tables_folder, 'rawdata')
     print(f"[INFO] Data summaries exported to {os.path.join(tables_folder, 'rawdata_info.csv')} and {os.path.join(tables_folder, 'rawdata_describe.csv')}")
