+++ scripts/02-clean.py
@@ -16,6 +16,14 @@ import pandas as pd
 import numpy as np
 import os
 import click
+from src.load_raw_data import load_raw_data
+from src.convert_categorical_to_numeric import convert_categorical_to_numeric
+from src.clean_text_header_columns import clean_text_header_columns
+from src.clean_encrypted_column import clean_encrypted_column
+from src.handle_numeric_error_codes import handle_numeric_error_codes
+from src.handle_categorical_nulls import handle_categorical_nulls
+from src.export_data_summary import export_data_summary
+from src.save_data import save_data
 
 # Default configuration constants
 DEFAULT_IMPORT_FILE_NAME = "PDFMalware2022.parquet"
@@ -23,216 +31,11 @@ DEFAULT_DATA_IMPORT_FOLDER = "../data/raw"
 DEFAULT_DATA_EXPORT_FOLDER = "../data/clean"
 DEFAULT_TABLES_FOLDER = "../results/tables"
 
-# Columns that require categorical-to-numeric conversion
-CATEGORICAL_NUMERIC_COLS = [
-    "Obj",
-    "Endobj",
-    "Xref",
-    "StartXref",
-    "PageNo",
-    "Images",
-    "JS",
-    "Javascript",
-    "AA",
-    "Endstream",
-    "OpenAction",
-    "Acroform",
-    "JBIG2Decode",
-    "RichMedia",
-    "Launch",
-    "EmbeddedFile",
-    "XFA",
-]
 
-# Text/header columns requiring special error code handling
-TEXT_HEADER_COLS = ["Text", "Header"]
+
 
 # Sentinel/error codes used in the raw data
 ERROR_CODES = [-1, 0]
-
-
-def load_raw_data(data_import_folder: str, import_file_name: str) -> pd.DataFrame:
-    """
-    Load raw parquet data from disk.
-
-    Args:
-        data_import_folder (str): Path to the folder containing the parquet file.
-        import_file_name (str): Name of the parquet file to load.
-
-    Returns:
-        pd.DataFrame: The raw dataset loaded from the parquet file.
-
-    Raises:
-        FileNotFoundError: If the specified parquet file does not exist.
-    """
-    data_path = os.path.join(data_import_folder, import_file_name)
-    raw_data = pd.read_parquet(data_path, engine="pyarrow")
-    return raw_data
-
-
-def convert_categorical_to_numeric(df: pd.DataFrame) -> None:
-    """
-    Convert categorical columns containing numeric strings to float type.
-
-    Extracts numeric values from categorical columns using regex, handling
-    negative numbers. Modifies the DataFrame in-place.
-
-    Args:
-        df (pd.DataFrame): The DataFrame to modify. Contains categorical columns
-                          with mixed string/numeric values.
-
-    Returns:
-        None
-    """
-    for col in CATEGORICAL_NUMERIC_COLS:
-        if col in df.columns:
-            # Convert to string, extract numeric pattern, convert to float
-            df[col] = (
-                df[col]
-                .astype(str)
-                .str.extract(r"(-?\d+)", expand=False)  # Extract numbers including negative
-                .astype("float")
-            )
-
-
-def clean_text_header_columns(df: pd.DataFrame) -> None:
-    """
-    Clean Text and Header columns by replacing error codes with NaN.
-
-    Converts columns to string type, marks error codes (0, -1) as NaN,
-    then converts to categorical dtype. Modifies the DataFrame in-place.
-
-    Args:
-        df (pd.DataFrame): The DataFrame to modify.
-
-    Returns:
-        None
-    """
-    for col in TEXT_HEADER_COLS:
-        if col in df.columns:
-            # Convert to string type
-            df[col] = df[col].astype("string")
-
-            # Replace error sentinel values with NaN
-            df.loc[df[col].isin(["0", "-1"]), col] = np.nan
-
-            # Convert to categorical for memory efficiency
-            df[col] = df[col].astype("category")
-
-
-def clean_encrypted_column(df: pd.DataFrame) -> None:
-    """
-    Clean the isEncrypted column: convert -1 to NaN, then binarize.
-
-    Replaces error code (-1) with NaN, then converts all values to binary
-    (0 or 1) where values > 0 are True. Modifies the DataFrame in-place.
-
-    Args:
-        df (pd.DataFrame): The DataFrame to modify. Must contain 'isEncrypted' column.
-
-    Returns:
-        None
-    """
-    if "isEncrypted" in df.columns:
-        # Replace error code with NaN
-        df.loc[df["isEncrypted"] == -1, "isEncrypted"] = np.nan
-
-        # Binarize: values > 0 become 1, others become 0
-        df["isEncrypted"] = (df["isEncrypted"] > 0).astype("float")
-
-
-def handle_numeric_error_codes(df: pd.DataFrame) -> None:
-    """
-    Replace error codes with NaN in all numeric columns.
-
-    Iterates through numeric columns and replaces -1 (common error sentinel)
-    with NaN. Modifies the DataFrame in-place.
-
-    Args:
-        df (pd.DataFrame): The DataFrame to modify. Contains numeric columns with -1 errors.
-
-    Returns:
-        None
-    """
-    # Identify all numeric columns
-    numeric_cols = [
-        col for col in df.columns
-        if df[col].dtype != "category" and col not in TEXT_HEADER_COLS
-    ]
-
-    # Replace error code -1 with NaN
-    for col in numeric_cols:
-        df.loc[df[col] == -1, col] = np.nan
-
-
-def handle_categorical_nulls(df: pd.DataFrame) -> None:
-    """
-    Normalize null handling in categorical columns.
-
-    Converts categorical columns to object type and ensures NaN values are
-    consistently represented. Modifies the DataFrame in-place.
-
-    Args:
-        df (pd.DataFrame): The DataFrame to modify. Contains categorical columns.
-
-    Returns:
-        None
-    """
-    # Identify categorical columns
-    cat_cols = [col for col in df.columns if df[col].dtype == "category"]
-
-    # Convert to object and normalize NaN representation
-    for col in cat_cols:
-        df[col] = df[col].astype(object)
-        df.loc[df[col].isna(), col] = np.nan
-
-
-def export_data_summary(clean_data: pd.DataFrame, tables_folder: str) -> None:
-    """
-    Export cleaned data information and statistical summaries to CSV files.
-
-    Generates two CSV files:
-        1. cleandata_info.csv: Column data types and null/non-null counts
-        2. cleandata_descr.csv: Statistical summaries (mean, std, quartiles, etc.)
-
-    Args:
-        clean_data (pd.DataFrame): The cleaned DataFrame to summarize.
-        tables_folder (str): Path to the folder where CSV files will be saved.
-
-    Returns:
-        None
-    """
-    # Export column-level information
-    info_df = pd.DataFrame(
-        {
-            "dtype": clean_data.dtypes,
-            "non_null_count": clean_data.notna().sum(),
-            "null_count": clean_data.isna().sum(),
-        }
-    )
-    info_path = os.path.join(tables_folder, "cleandata_info.csv")
-    info_df.to_csv(info_path)
-
-    # Export statistical summaries
-    descr_df = clean_data.describe()
-    descr_path = os.path.join(tables_folder, "cleandata_descr.csv")
-    descr_df.to_csv(descr_path)
-
-
-def save_clean_data(clean_data: pd.DataFrame, data_export_folder: str) -> None:
-    """
-    Save cleaned data to CSV file.
-
-    Args:
-        clean_data (pd.DataFrame): The cleaned DataFrame to save.
-        data_export_folder (str): Path to the folder where the CSV will be saved.
-
-    Returns:
-        None
-    """
-    os.makedirs(data_export_folder, exist_ok=True)
-    clean_data_path = os.path.join(data_export_folder, "cleandata.csv")
-    clean_data.to_csv(clean_data_path, index=False)
 
 
 @click.command()
@@ -302,10 +105,10 @@ def main(import_file_name: str, data_import_folder: str, data_export_folder: str
 
 
     # Step 7: Export summaries
-    export_data_summary(clean_data, tables_folder)
+    export_data_summary(clean_data, tables_folder, 'cleandata')
     print(f"[INFO] Data summaries exported to {os.path.join(tables_folder, 'cleandata_info.csv')} and {os.path.join(tables_folder, 'cleandata_descr.csv')}")
     # Step 8: Save cleaned data
-    save_clean_data(clean_data, data_export_folder)
+    save_data(clean_data, data_export_folder, 'cleandata')
     print(f"[INFO] Cleaned data saved to {os.path.join(data_export_folder, 'cleandata.csv')}")
 if __name__ == "__main__":
     main()
