+++ scripts/03-validate.py
@@ -12,6 +12,7 @@ The validation pipeline ensures data quality and consistency before
 downstream analysis and model training.
 """
 
+
 import pandera.pandas as pa
 import pandas as pd
 import numpy as np
@@ -19,447 +20,13 @@ import json
 import logging
 from scipy import stats
 import click
+from src.build_validation_schema import build_validation_schema
+from src.validate_class_balance import validate_class_balance
+from src.validate_schema import validate_schema
+from src.check_outliers import check_outliers
 
 # Default configuration constants
 DEFAULT_CLEAN_DATA_PATH = "../data/clean/cleandata.csv"
-
-# Validation parameters
-NULLNESS_THRESHOLD = 0.05  # Maximum proportion of null values allowed (5%)
-CLASS_BALANCE_THRESHOLD = 0.15  # Maximum class imbalance (15% difference)
-OUTLIER_Z_SCORE = 3  # Z-score threshold for outlier detection
-
-
-def build_validation_schema() -> pa.DataFrameSchema:
-    """
-    Build and return the data validation schema.
-
-    Defines the expected structure of the cleaned PDF malware dataset including
-    column names, data types, nullness constraints, and value ranges.
-
-    Returns:
-        pa.DataFrameSchema: Pandera schema object defining validation rules.
-    """
-    # Schema definition with column constraints and validations
-    schema = pa.DataFrameSchema(
-        {
-            "FileName": pa.Column(str, nullable=False),
-            "PdfSize": pa.Column(
-                float,
-                checks=[
-                    pa.Check.ge(0),
-                    pa.Check(
-                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                        element_wise=False,
-                        error=f"Too many null values in 'PdfSize' column.",
-                    ),
-                ],
-                nullable=True,
-                coerce=True,
-            ),
-            "MetadataSize": pa.Column(
-                float,
-                checks=[
-                    pa.Check.ge(0),
-                    pa.Check(
-                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                        element_wise=False,
-                        error=f"Too many null values in 'MetadataSize' column.",
-                    ),
-                ],
-                nullable=True,
-                coerce=True,
-            ),
-            "Pages": pa.Column(
-                float,
-                checks=[
-                    pa.Check.ge(0),
-                    pa.Check(
-                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                        element_wise=False,
-                        error=f"Too many null values in 'Pages' column.",
-                    ),
-                ],
-                nullable=True,
-                coerce=True,
-            ),
-            "XrefLength": pa.Column(
-                float,
-                checks=[
-                    pa.Check.ge(0),
-                    pa.Check(
-                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                        element_wise=False,
-                        error=f"Too many null values in 'XrefLength' column.",
-                    ),
-                ],
-                nullable=True,
-                coerce=True,
-            ),
-            "TitleCharacters": pa.Column(
-                float,
-                checks=[
-                    pa.Check.ge(0),
-                    pa.Check(
-                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                        element_wise=False,
-                        error=f"Too many null values in 'TitleCharacters' column.",
-                    ),
-                ],
-                nullable=True,
-                coerce=True,
-            ),
-            "isEncrypted": pa.Column(float, nullable=False),
-            "EmbeddedFiles": pa.Column(
-                float,
-                checks=[
-                    pa.Check.ge(0),
-                    pa.Check(
-                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                        element_wise=False,
-                        error=f"Too many null values in 'EmbeddedFiles' column.",
-                    ),
-                ],
-                nullable=True,
-                coerce=True,
-            ),
-            "Images": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'Images' column.",
-                ),
-                nullable=True,
-            ),
-            "Text": pa.Column(
-                str,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'Text' column.",
-                ),
-                nullable=True,
-            ),
-            "Header": pa.Column(
-                str,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'Header' column.",
-                ),
-                nullable=True,
-            ),
-            "Obj": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'Obj' column.",
-                ),
-                nullable=True,
-            ),
-            "Endobj": pa.Column(
-                float,
-                checks=[
-                    pa.Check.ge(0),
-                    pa.Check(
-                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                        element_wise=False,
-                        error=f"Too many null values in 'Endobj' column.",
-                    ),
-                ],
-                nullable=True,
-                coerce=True,
-            ),
-            "Stream": pa.Column(
-                float,
-                checks=[
-                    pa.Check.ge(0),
-                    pa.Check(
-                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                        element_wise=False,
-                        error=f"Too many null values in 'Stream' column.",
-                    ),
-                ],
-                nullable=True,
-                coerce=True,
-            ),
-            "Endstream": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'Endstream' column.",
-                ),
-                nullable=True,
-            ),
-            "Xref": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'Xref' column.",
-                ),
-                nullable=True,
-            ),
-            "Trailer": pa.Column(
-                float,
-                checks=[
-                    pa.Check.ge(0),
-                    pa.Check(
-                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                        element_wise=False,
-                        error=f"Too many null values in 'Trailer' column.",
-                    ),
-                ],
-                nullable=True,
-                coerce=True,
-            ),
-            "StartXref": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'StartXref' column.",
-                ),
-                nullable=True,
-            ),
-            "PageNo": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'PageNo' column.",
-                ),
-                nullable=True,
-            ),
-            "Encrypt": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'Encrypt' column.",
-                ),
-                nullable=True,
-                coerce=True,
-            ),
-            "ObjStm": pa.Column(
-                float,
-                checks=[
-                    pa.Check.ge(0),
-                    pa.Check(
-                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                        element_wise=False,
-                        error=f"Too many null values in 'ObjStm' column.",
-                    ),
-                ],
-                nullable=True,
-                coerce=True,
-            ),
-            "JS": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'JS' column.",
-                ),
-                nullable=True,
-            ),
-            "Javascript": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'Javascript' column.",
-                ),
-                nullable=True,
-            ),
-            "AA": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'AA' column.",
-                ),
-                nullable=True,
-            ),
-            "OpenAction": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'OpenAction' column.",
-                ),
-                nullable=True,
-            ),
-            "Acroform": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'Acroform' column.",
-                ),
-                nullable=True,
-            ),
-            "JBIG2Decode": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'JBIG2Decode' column.",
-                ),
-                nullable=True,
-            ),
-            "RichMedia": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'RichMedia' column.",
-                ),
-                nullable=True,
-            ),
-            "Launch": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'Launch' column.",
-                ),
-                nullable=True,
-            ),
-            "EmbeddedFile": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'EmbeddedFile' column.",
-                ),
-                nullable=True,
-            ),
-            "XFA": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'XFA' column.",
-                ),
-                nullable=True,
-            ),
-            "Colors": pa.Column(
-                float,
-                pa.Check(
-                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
-                    element_wise=False,
-                    error=f"Too many null values in 'Colors' column.",
-                ),
-                nullable=True,
-                coerce=True,
-            ),
-            "Class": pa.Column(
-                str,
-                pa.Check.isin(["Malicious", "Benign"]),
-                nullable=False,
-            ),
-        },
-        checks=[
-            pa.Check(
-                lambda clean_data: ~clean_data.duplicated().any(),
-                error="Duplicate rows found.",
-            ),
-            pa.Check(
-                lambda clean_data: ~(clean_data.isna().all(axis=1)).any(),
-                error="Empty rows found.",
-            ),
-        ],
-    )
-    return schema
-
-
-def validate_schema(clean_data: pd.DataFrame, schema: pa.DataFrameSchema) -> bool:
-    """
-    Validate data against the defined schema.
-
-    Performs schema validation including column names, data types, constraints,
-    duplicates, and empty rows. Logs validation errors if any are found.
-
-    Args:
-        clean_data (pd.DataFrame): The DataFrame to validate.
-        schema (pa.DataFrameSchema): The validation schema to apply.
-
-    Returns:
-        bool: True if validation passes, False otherwise.
-    """
-    try:
-        schema.validate(clean_data, lazy=True)
-        print("All data validation checks are passed!")
-        return True
-    except pa.errors.SchemaErrors as e:
-        print("Data validation checks failed, please see above")
-        error_message = json.dumps(e.message, indent=2)
-        logging.error("\n" + error_message)
-        return False
-
-
-def check_outliers(clean_data: pd.DataFrame) -> bool:
-    """
-    Detect outliers using Z-score analysis on numeric columns.
-
-    Identifies anomalous values by computing Z-scores and flagging values
-    exceeding the threshold. Warns user if outliers are detected.
-
-    Args:
-        clean_data (pd.DataFrame): The DataFrame to check for outliers.
-
-    Returns:
-        bool: True if outliers are detected, False otherwise.
-    """
-    # Calculate Z-scores for numeric columns, excluding NaN values
-    numeric_data = clean_data.dropna().select_dtypes(include=np.number)
-
-    if numeric_data.empty:
-        print("No numeric data available for outlier detection")
-        return False
-
-    z_scores = np.abs(stats.zscore(numeric_data))
-    outliers = (z_scores > OUTLIER_Z_SCORE)
-
-    if outliers.any():
-        print(f"Warning: outliers detected (Z-score > {OUTLIER_Z_SCORE})")
-        return True
-
-    return False
-
-
-def validate_class_balance(clean_data: pd.DataFrame) -> bool:
-    """
-    Validate that target class distribution meets balance requirements.
-
-    Checks that the difference in proportions between the two classes
-    does not exceed the defined threshold (default 15%).
-
-    Args:
-        clean_data (pd.DataFrame): The DataFrame containing the 'Class' column.
-
-    Returns:
-        bool: True if balance is acceptable, False if threshold is exceeded.
-    """
-    # Get class distribution as normalized proportions
-    clean_data_balance = clean_data["Class"].value_counts(normalize=True)
-
-    # Calculate difference between class proportions
-    class_diff = np.abs(clean_data_balance.iloc[0] - clean_data_balance.iloc[1])
-
-    if class_diff >= CLASS_BALANCE_THRESHOLD:
-        print(
-            f"Target variable distribution check failed. "
-            f"The count difference between classes ({class_diff:.2%}) "
-            f"exceeds the threshold ({CLASS_BALANCE_THRESHOLD:.2%})."
-        )
-        return False
-
-    return True
-
 
 @click.command()
 @click.option(
@@ -468,6 +35,7 @@ def validate_class_balance(clean_data: pd.DataFrame) -> bool:
     show_default=True,
     help="Path to the cleaned data CSV to validate.",
 )
+
 def main(clean_data_path: str) -> None:
     """
     Main data validation pipeline.
@@ -499,5 +67,6 @@ def main(clean_data_path: str) -> None:
     # Step 4: Validate class balance
     balance_valid = validate_class_balance(clean_data)
     print(f"Data validation pipeline completed successfully: {schema_valid and has_outliers and balance_valid}")
+
 if __name__ == "__main__":
     main()
