+++ scripts/04-datasplit.py
@@ -13,104 +13,16 @@ import os
 import click
 import pandas as pd
 from sklearn.model_selection import train_test_split
+from src.load_data import load_data
+from src.extract_features import extract_features
+from src.split_data import split_data
+from src.save_data import save_data
 
 # Default configuration constants
 DEFAULT_CLEAN_DATA_PATH = "../data/clean/cleandata.csv"
 DEFAULT_SPLIT_SAVE_FOLDER = "../data/split"
 
-# Dataset splitting parameters
-TEST_SPLIT_SIZE = 0.2  # 80/20 train/test split
-RANDOM_STATE = 42  # Fixed seed for reproducibility
 
-
-def load_data(clean_data_path: str) -> pd.DataFrame:
-    """
-    Load cleaned data from CSV file.
-
-    Args:
-        clean_data_path (str): Path to the cleaned data CSV file.
-
-    Returns:
-        pd.DataFrame: The loaded DataFrame.
-
-    Raises:
-        FileNotFoundError: If the specified file does not exist.
-    """
-    clean_data = pd.read_csv(clean_data_path)
-    return clean_data
-
-
-def extract_features(data: pd.DataFrame) -> pd.DataFrame:
-    """
-    Extract features by removing non-predictive identifier columns.
-
-    Removes columns like 'FileName' that are identifiers rather than features.
-    If FileName is not present, returns the data unchanged.
-
-    Args:
-        data (pd.DataFrame): The full dataset with all columns.
-
-    Returns:
-        pd.DataFrame: DataFrame with identifier columns removed.
-    """
-    if "FileName" in data.columns:
-        # Remove identifier column; keep all features and target
-        features = data.drop("FileName", axis=1)
-    else:
-        # No identifier columns to remove
-        features = data
-
-    return features
-
-
-def split_data(features: pd.DataFrame) -> tuple[pd.DataFrame, pd.DataFrame]:
-    """
-    Perform stratified train/test split on the feature data.
-
-    Splits data with 80% for training and 20% for testing using a fixed
-    random state to ensure reproducibility.
-
-    Args:
-        features (pd.DataFrame): The feature DataFrame to split. Should include
-                                the target column 'Class' if present.
-
-    Returns:
-        tuple[pd.DataFrame, pd.DataFrame]: A tuple containing:
-            - train: Training data (80%)
-            - test: Test data (20%)
-    """
-    train, test = train_test_split(
-        features,
-        test_size=TEST_SPLIT_SIZE,
-        random_state=RANDOM_STATE
-    )
-    return train, test
-
-
-def save_split_data(train: pd.DataFrame, test: pd.DataFrame, split_data_path: str) -> None:
-    """
-    Save train and test splits to separate CSV files.
-
-    Creates the output directory if it doesn't exist, then saves both
-    train and test DataFrames as CSV files without index.
-
-    Args:
-        train (pd.DataFrame): The training dataset.
-        test (pd.DataFrame): The test dataset.
-        split_data_path (str): Path to the folder where CSV files will be saved.
-
-    Returns:
-        None
-    """
-    # Create output directory if needed
-    os.makedirs(split_data_path, exist_ok=True)
-
-    # Save train and test splits
-    train_path = os.path.join(split_data_path, "traindata.csv")
-    test_path = os.path.join(split_data_path, "testdata.csv")
-
-    train.to_csv(train_path, index=False)
-    test.to_csv(test_path, index=False)
 
 
 @click.command()
@@ -154,8 +66,9 @@ def main(clean_data_path: str, split_data_path: str) -> None:
     train, test = split_data(features)
     print(f"[INFO] Data split into train and test successfully")
     # Step 4: Save split data
-    save_split_data(train, test, split_data_path)
-    print(f"[INFO] Split data saved to {os.path.join(split_data_path, 'traindata.csv')} and {os.path.join(split_data_path, 'testdata.csv')}")
+    save_data(train, split_data_path, 'traindata')
+    save_data(test, split_data_path, 'testdata')
+    print(f"[INFO] train and test data saved to {os.path.join(split_data_path, 'traindata.csv')} and {os.path.join(split_data_path, 'testdata.csv')}")
 
 if __name__ == "__main__":
     main()
