+++ scripts/05-eda.py
@@ -21,6 +21,13 @@ import altair as alt
 import altair_ally as aly
 import matplotlib.pyplot as plt
 import warnings
+from src.load_data import load_data
+from src.compute_shape_stats import compute_shape_stats
+from src.compute_dtype_distribution import compute_dtype_distribution
+from src.analyze_class_balance import analyze_class_balance
+from src.generate_descriptive_stats import generate_descriptive_stats
+from src.generate_distribution_charts import generate_distribution_charts
+from src.generate_correlation_chart import generate_correlation_chart
 
 # Suppress packages throwing FutureWarning & AltairDeprecationWarning terminal messages
 warnings.filterwarnings("ignore")
@@ -29,157 +36,6 @@ warnings.filterwarnings("ignore")
 DEFAULT_TRAIN_PATH = "../data/split/traindata.csv"
 DEFAULT_CHARTS_FOLDER = "../results/charts"
 DEFAULT_TABLES_FOLDER = "../results/tables"
-
-
-def load_training_data(train_path: str) -> pd.DataFrame:
-    """
-    Load training data from CSV file.
-
-    Args:
-        train_path (str): Path to the training data CSV file.
-
-    Returns:
-        pd.DataFrame: The loaded training dataset.
-
-    Raises:
-        FileNotFoundError: If the specified file does not exist.
-    """
-    train = pd.read_csv(train_path)
-    return train
-
-
-def compute_shape_stats(train: pd.DataFrame, tables_folder: str) -> None:
-    """
-    Compute and save dataset shape statistics.
-
-    Calculates the number of rows and columns in the training dataset
-    and saves to a CSV file.
-
-    Args:
-        train (pd.DataFrame): The training dataset.
-        tables_folder (str): Path to the folder where CSV will be saved.
-
-    Returns:
-        None
-    """
-    shape_df = pd.DataFrame({
-        "rows": [train.shape[0]],
-        "columns": [train.shape[1]]
-    })
-    shape_path = os.path.join(tables_folder, "train_shape.csv")
-    shape_df.to_csv(shape_path, index=False)
-
-
-def compute_dtype_distribution(train: pd.DataFrame, tables_folder: str) -> None:
-    """
-    Compute and save data type distribution.
-
-    Analyzes the frequency of each data type (float, object, category, etc.)
-    in the dataset and saves to a CSV file.
-
-    Args:
-        train (pd.DataFrame): The training dataset.
-        tables_folder (str): Path to the folder where CSV will be saved.
-
-    Returns:
-        None
-    """
-    # Count occurrences of each data type
-    dtype_counts = train.dtypes.value_counts().reset_index()
-    dtype_counts.columns = ["Data Type", "Count"]
-
-    dtype_path = os.path.join(tables_folder, "dtype_counts.csv")
-    dtype_counts.to_csv(dtype_path, index=False)
-
-
-def analyze_class_balance(train: pd.DataFrame, tables_folder: str, charts_folder: str) -> None:
-    """
-    Analyze and visualize target class balance.
-
-    Generates both a CSV table with class proportions and a bar chart
-    visualization showing the distribution of the target variable.
-
-    Args:
-        train (pd.DataFrame): The training dataset (must contain 'Class' column).
-        tables_folder (str): Path to folder where CSV will be saved.
-        charts_folder (str): Path to folder where PNG chart will be saved.
-
-    Returns:
-        None
-    """
-    # Calculate class proportions
-    train_balance = train["Class"].value_counts(normalize=True).round(4)
-    balance_path_table = os.path.join(tables_folder, "train_class_balance.csv")
-    train_balance.to_csv(balance_path_table, header=["Proportion"])
-
-    # Generate and save bar chart
-    counts = train["Class"].value_counts()
-    counts.plot(kind="bar", title="Class Balance", rot=0)
-    balance_path_chart = os.path.join(charts_folder, "train_class_balance.png")
-    plt.tight_layout()
-    plt.savefig(balance_path_chart)
-    plt.close()
-
-
-def generate_descriptive_stats(train: pd.DataFrame, tables_folder: str) -> None:
-    """
-    Generate and save descriptive statistics for all features.
-
-    Computes statistical summaries (mean, std, min, max, quartiles, etc.)
-    for all columns in the dataset.
-
-    Args:
-        train (pd.DataFrame): The training dataset.
-        tables_folder (str): Path to folder where CSV will be saved.
-
-    Returns:
-        None
-    """
-    train_descr_path = os.path.join(tables_folder, "train_describe.csv")
-    train.describe(include="all").round(2).to_csv(train_descr_path)
-
-
-def generate_distribution_charts(train: pd.DataFrame, charts_folder: str) -> None:
-    """
-    Generate univariate distribution charts for all features.
-
-    Creates interactive HTML charts showing the distribution of each feature,
-    colored by the target class.
-
-    Args:
-        train (pd.DataFrame): The training dataset (must contain 'Class' column).
-        charts_folder (str): Path to folder where HTML chart will be saved.
-
-    Returns:
-        None
-    """
-    # Disable row limit for Altair to handle large datasets
-    alt.data_transformers.disable_max_rows()
-
-    # Generate univariate distribution chart with 4 columns layout
-    univar_chart = aly.dist(train, color='Class', columns=4)
-    univar_path = os.path.join(charts_folder, "train_univar_distribution.html")
-    univar_chart.save(univar_path)
-
-
-def generate_correlation_chart(train: pd.DataFrame, charts_folder: str) -> None:
-    """
-    Generate correlation matrix visualization.
-
-    Creates an interactive HTML heatmap showing pairwise correlations
-    between all numeric features.
-
-    Args:
-        train (pd.DataFrame): The training dataset with numeric features.
-        charts_folder (str): Path to folder where HTML chart will be saved.
-
-    Returns:
-        None
-    """
-    corr_chart = aly.corr(train)
-    corr_path = os.path.join(charts_folder, "train_correlations.html")
-    corr_chart.save(corr_path)
-
 
 @click.command()
 @click.option(
@@ -227,7 +83,7 @@ def main(train_path: str, tables_folder: str, charts_folder: str) -> None:
 
     print(f"[INFO] Starting EDA pipeline with train path: {train_path}, tables folder: {tables_folder}, and charts folder: {charts_folder}")
     # Step 1: Load training data
-    train = load_training_data(train_path)
+    train = load_data(train_path)
     print(f"[INFO] Completed step 1/7")
     # Step 2: Dataset shape statistics
     compute_shape_stats(train, tables_folder)
