+++ scripts/06-preprocessing.py
@@ -20,238 +20,18 @@ from sklearn.pipeline import Pipeline, make_pipeline
 from sklearn.impute import SimpleImputer
 from sklearn.preprocessing import StandardScaler, OneHotEncoder
 from typing import List
+from src.infer_feature_groups import infer_feature_groups
+from src. build_preprocessor import build_preprocessor
+from src.load_data import load_data
+from src.extract_features_and_labels import extract_features_and_labels
+from src.fit_and_transform_data import fit_and_transform_data
+from src.create_output_dataframes import create_output_dataframes
+from src.save_data import save_data
 
 # Default configuration constants
 DEFAULT_TRAIN_PATH = "../data/split/traindata.csv"
 DEFAULT_TEST_PATH = "../data/split/testdata.csv"
 DEFAULT_OUTPUT_FOLDER = "../data/transformed"
-
-# Default target column name
-TARGET_COLUMN = "Class"
-
-
-def infer_feature_groups(df: pd.DataFrame, target_col: str = TARGET_COLUMN) -> tuple[list[str], list[str]]:
-    """
-    Automatically infer numeric and categorical feature columns.
-
-    Separates features into numeric and categorical types, excluding the
-    target column. Uses pandas dtype inference to classify columns.
-
-    Args:
-        df (pd.DataFrame): The DataFrame to analyze for feature types.
-        target_col (str): Name of the target column to exclude (default: "Class").
-
-    Returns:
-        tuple[list[str], list[str]]: A tuple containing:
-            - numeric column names
-            - categorical column names (object or category dtype)
-    """
-    # Remove target column for feature analysis
-    feature_df = df.drop(columns=[target_col], errors="ignore")
-
-    # Identify categorical columns (object or category dtype)
-    cat_cols = feature_df.select_dtypes(include=["object", "category"]).columns.tolist()
-
-    # Numeric columns are everything else
-    num_cols = feature_df.columns.difference(cat_cols).tolist()
-
-    return num_cols, cat_cols
-
-
-def build_preprocessor(numeric_cols: List[str], cat_cols: List[str]) -> ColumnTransformer:
-    """
-    Build a preprocessing pipeline for numeric and categorical features.
-
-    Creates a ColumnTransformer that applies different preprocessing steps
-    to numeric and categorical columns:
-    - Numeric: imputation (median) then standardization
-    - Categorical: imputation (mode) then one-hot encoding
-
-    Args:
-        numeric_cols (List[str]): List of numeric column names.
-        cat_cols (List[str]): List of categorical column names.
-
-    Returns:
-        ColumnTransformer: Sklearn preprocessor with separate pipelines
-                          for numeric and categorical features.
-    """
-    # Pipeline for numeric features: handle missing values then standardize
-    numeric_transformer = Pipeline(
-        steps=[
-            ("imputer", SimpleImputer(strategy="median")),  # Use median for robustness
-            ("scaler", StandardScaler()),  # Zero mean, unit variance
-        ]
-    )
-
-    # Pipeline for categorical features: handle missing values then encode
-    categorical_transformer = Pipeline(
-        steps=[
-            ("imputer", SimpleImputer(strategy="most_frequent")),  # Use mode for categories
-            ("onehot", OneHotEncoder(handle_unknown="ignore")),  # Convert to numeric
-        ]
-    )
-
-    # Combine both pipelines using ColumnTransformer
-    preprocessor = ColumnTransformer(
-        transformers=[
-            ("num", numeric_transformer, numeric_cols),
-            ("cat", categorical_transformer, cat_cols),
-        ]
-    )
-
-    return preprocessor
-
-
-def load_data(train_path: str, test_path: str) -> tuple[pd.DataFrame, pd.DataFrame]:
-    """
-    Load training and test data from CSV files.
-
-    Args:
-        train_path (str): Path to the training data CSV.
-        test_path (str): Path to the test data CSV.
-
-    Returns:
-        tuple[pd.DataFrame, pd.DataFrame]: A tuple containing:
-            - train_df: The training DataFrame
-            - test_df: The test DataFrame
-
-    Raises:
-        FileNotFoundError: If either file does not exist.
-    """
-    train_df = pd.read_csv(train_path)
-    test_df = pd.read_csv(test_path)
-    return train_df, test_df
-
-
-def extract_features_and_labels(
-    train_df: pd.DataFrame,
-    test_df: pd.DataFrame,
-    target_col: str = TARGET_COLUMN
-) -> tuple[pd.DataFrame, pd.Series | None, pd.DataFrame, pd.Series | None]:
-    """
-    Extract features and labels from train/test DataFrames.
-
-    Separates the target variable from features. If the target column is absent,
-    returns None for the labels.
-
-    Args:
-        train_df (pd.DataFrame): The training DataFrame.
-        test_df (pd.DataFrame): The test DataFrame.
-        target_col (str): Name of the target column (default: "Class").
-
-    Returns:
-        tuple: A tuple containing:
-            - X_train: Training features
-            - y_train: Training labels (or None if not present)
-            - X_test: Test features
-            - y_test: Test labels (or None if not present)
-    """
-    # Extract features and labels from training data
-    X_train = train_df.drop(columns=[target_col], errors="ignore")
-    y_train = train_df[target_col] if target_col in train_df.columns else None
-
-    # Extract features and labels from test data
-    X_test = test_df.drop(columns=[target_col], errors="ignore")
-    y_test = test_df[target_col] if target_col in test_df.columns else None
-
-    return X_train, y_train, X_test, y_test
-
-
-def fit_and_transform_data(
-    preprocessor: ColumnTransformer,
-    X_train: pd.DataFrame,
-    X_test: pd.DataFrame
-) -> tuple[pd.DataFrame, pd.DataFrame]:
-    """
-    Fit preprocessor on training data and transform both train and test sets.
-
-    Applies the train-fit pattern: learns statistics from training data
-    and applies the same transformation to both training and test sets.
-
-    Args:
-        preprocessor (ColumnTransformer): The preprocessing pipeline.
-        X_train (pd.DataFrame): Training features.
-        X_test (pd.DataFrame): Test features.
-
-    Returns:
-        tuple[pd.DataFrame, pd.DataFrame]: A tuple containing:
-            - X_train_transformed: Preprocessed training features
-            - X_test_transformed: Preprocessed test features
-    """
-    # Fit preprocessor on training data
-    preprocessor.fit(X_train)
-
-    # Transform both training and test sets
-    X_train_t = preprocessor.transform(X_train)
-    X_test_t = preprocessor.transform(X_test)
-
-    return X_train_t, X_test_t
-
-
-def create_output_dataframes(
-    X_train_t: pd.DataFrame | None,
-    X_test_t: pd.DataFrame | None,
-    preprocessor: ColumnTransformer,
-    y_train: pd.Series | None,
-    y_test: pd.Series | None,
-    target_col: str = TARGET_COLUMN
-) -> tuple[pd.DataFrame, pd.DataFrame]:
-    """
-    Create output DataFrames from transformed features and original labels.
-
-    Converts transformed numpy arrays back to DataFrames with proper column names
-    and optionally appends the target variable.
-
-    Args:
-        X_train_t: Transformed training features (numpy array or sparse matrix).
-        X_test_t: Transformed test features (numpy array or sparse matrix).
-        preprocessor (ColumnTransformer): Fitted preprocessor with feature names.
-        y_train (pd.Series | None): Original training labels.
-        y_test (pd.Series | None): Original test labels.
-        target_col (str): Name of the target column (default: "Class").
-
-    Returns:
-        tuple[pd.DataFrame, pd.DataFrame]: A tuple containing:
-            - train_out: Training DataFrame with transformed features and labels
-            - test_out: Test DataFrame with transformed features and labels
-    """
-    # Get feature names from the fitted preprocessor
-    feature_names = preprocessor.get_feature_names_out()
-
-    # Convert transformed arrays to DataFrames
-    train_out = pd.DataFrame(X_train_t, columns=feature_names)
-    test_out = pd.DataFrame(X_test_t, columns=feature_names)
-
-    # Append target variable if present
-    if y_train is not None:
-        train_out[target_col] = y_train.reset_index(drop=True)
-
-    if y_test is not None:
-        test_out[target_col] = y_test.reset_index(drop=True)
-
-    return train_out, test_out
-
-
-def save_transformed_data(
-    train_out: pd.DataFrame,
-    test_out: pd.DataFrame,
-    output_folder: str
-) -> None:
-    """
-    Save transformed training and test data to CSV files.
-
-    Args:
-        train_out (pd.DataFrame): Transformed training DataFrame.
-        test_out (pd.DataFrame): Transformed test DataFrame.
-        output_folder (str): Path to folder where CSV files will be saved.
-
-    Returns:
-        None
-    """
-    os.makedirs(output_folder, exist_ok=True)
-
-    train_out.to_csv(os.path.join(output_folder, "train_transformed.csv"), index=False)
-    test_out.to_csv(os.path.join(output_folder, "test_transformed.csv"), index=False)
 
 
 @click.command()
@@ -315,7 +95,8 @@ def main(
     """
     print(f"[INFO] Starting preprocessing pipeline with train path: {train_path}, test path: {test_path}, output folder: {output_folder}")
     # Step 1: Load data
-    train_df, test_df = load_data(train_path, test_path)
+    train_df = load_data(train_path)
+    test_df = load_data(test_path)
     print(f"[INFO] Completed step 1/7")
     # Step 2: Determine feature groups (auto-infer or use provided)
     if numeric_col or cat_col:
@@ -339,7 +120,8 @@ def main(
     )
     print(f"[INFO] Completed step 6/7")
     # Step 7: Save transformed data
-    save_transformed_data(train_out, test_out, output_folder)
+    save_data(train_out, output_folder, "train_transformed")
+    save_data(test_out, output_folder, "test_transformed")
     print(f"[INFO] Completed step 7/7")
     print(f"[INFO] Preprocessing pipeline completed successfully. Transformed data saved to {os.path.join(output_folder, 'train_transformed.csv')} and {os.path.join(output_folder, 'test_transformed.csv')}")
