+++ scripts/07-train.py
@@ -25,261 +25,18 @@ from sklearn.linear_model import LogisticRegression
 from sklearn.svm import SVC
 from scipy.stats import loguniform
 import joblib
+from src.get_model_and_search_space import get_model_and_search_space
+from src.load_split_data import load_split_data
+from src.define_scoring_metrics import define_scoring_metrics
+from src.perform_cross_validation import perform_cross_validation
+from src.perform_hyperparameter_optimization import perform_hyperparameter_optimization
+from src.save_model import save_model
+from src.save_cv_results import save_cv_results
 
 # Default configuration constants
 DEFAULT_TRAIN_PATH = "../data/transformed/train_transformed.csv"
 DEFAULT_CV_OUTPUT_FOLDER = "../results/tables"
 DEFAULT_MODELS_FOLDER = "../results/models"
-
-# Model training parameters
-CV_FOLDS = 5  # Number of cross-validation folds
-HYPERPARAMETER_ITERATIONS = 5  # Number of hyperparameter combinations to try
-RANDOM_STATE = 42  # Fixed seed for reproducibility
-TARGET_POSITIVE_LABEL = "Malicious"  # Positive class for metrics
-
-
-def get_model_and_search_space(model_type: str):
-    """
-    Get model instance and hyperparameter search space for a given model type.
-
-    Returns a scikit-learn estimator and corresponding hyperparameter search space
-    for use with RandomizedSearchCV. Each model has different hyperparameters
-    and search ranges.
-
-    Args:
-        model_type (str): Type of model to instantiate. Options:
-            - "dummy": Dummy classifier baseline
-            - "logisticreg": Logistic regression
-            - "svc": Support vector classifier
-
-    Returns:
-        tuple: A tuple containing:
-            - model: Instantiated scikit-learn estimator
-            - search_space (dict): Hyperparameter distributions for RandomizedSearchCV
-
-    Raises:
-        ValueError: If model_type is not supported.
-    """
-    if model_type == "dummy":
-        # Baseline classifier that makes simple predictions
-        model = DummyClassifier(random_state=RANDOM_STATE)
-        search_space = {
-            "strategy": ["most_frequent", "prior"]  # Baseline strategies
-        }
-
-    elif model_type == "logisticreg":
-        # Logistic regression classifier
-        model = LogisticRegression(max_iter=1000, random_state=RANDOM_STATE)
-        search_space = {
-            "C": loguniform(1e-3, 1e3),  # Inverse regularization strength
-            "class_weight": [None, "balanced"],  # Handle class imbalance
-        }
-
-    elif model_type == "svc":
-        # Support Vector Classifier
-        model = SVC(random_state=RANDOM_STATE)
-        search_space = {
-            "C": loguniform(1e-3, 1e3),  # Regularization parameter
-            "gamma": loguniform(1e-4, 1e1),  # Kernel coefficient
-            "class_weight": [None, "balanced"],  # Handle class imbalance
-        }
-
-    else:
-        raise ValueError(
-            f"Unsupported model type: {model_type}. "
-            f"Supported types: 'dummy', 'logisticreg', 'svc'"
-        )
-
-    return model, search_space
-
-
-def load_training_data(train_path: str) -> tuple[pd.DataFrame, pd.Series]:
-    """
-    Load training data and extract features and labels.
-
-    Reads training data from CSV and validates that the required 'Class'
-    column is present.
-
-    Args:
-        train_path (str): Path to the training CSV file.
-
-    Returns:
-        tuple[pd.DataFrame, pd.Series]: A tuple containing:
-            - X_train: Training features (all columns except 'Class')
-            - y_train: Training labels (the 'Class' column)
-
-    Raises:
-        ValueError: If 'Class' column is missing from the data.
-    """
-    print(f"[INFO] Loading training data")
-    train_df = pd.read_csv(train_path)
-
-    if "Class" not in train_df.columns:
-        raise ValueError("Training data must include a 'Class' column.")
-
-    X_train = train_df.drop(columns=["Class"])
-    y_train = train_df["Class"]
-
-    print(f"[INFO] Loaded training data. Rows: {len(train_df)}, Features: {X_train.shape[1]}")
-
-    return X_train, y_train
-
-
-def define_scoring_metrics() -> dict:
-    """
-    Define evaluation metrics for model cross-validation.
-
-    Creates a dictionary of scoring functions for evaluating model performance,
-    with emphasis on precision and recall for the "Malicious" class.
-
-    Returns:
-        dict: Dictionary mapping metric names to scoring functions or scorer objects.
-    """
-    scoring = {
-        "accuracy": "accuracy",  # Proportion of correct predictions
-        "precision": make_scorer(
-            precision_score,
-            zero_division=0,
-            pos_label=TARGET_POSITIVE_LABEL
-        ),  # True positives / all positive predictions
-        "recall": make_scorer(
-            recall_score,
-            zero_division=0,
-            pos_label=TARGET_POSITIVE_LABEL
-        ),  # True positives / all actual positives
-        "f1": make_scorer(
-            f1_score,
-            zero_division=0,
-            pos_label=TARGET_POSITIVE_LABEL
-        ),  # Harmonic mean of precision and recall
-    }
-    return scoring
-
-
-def perform_cross_validation(
-    model,
-    X_train: pd.DataFrame,
-    y_train: pd.Series,
-    scoring: dict,
-    cv_folds: int = CV_FOLDS
-) -> pd.DataFrame:
-    """
-    Perform cross-validation and return aggregated results.
-
-    Evaluates model performance using k-fold cross-validation, computing
-    both training and validation scores for each metric.
-
-    Args:
-        model: Scikit-learn estimator to evaluate.
-        X_train (pd.DataFrame): Training features.
-        y_train (pd.Series): Training labels.
-        scoring (dict): Dictionary of scoring metrics.
-        cv_folds (int): Number of cross-validation folds (default: 5).
-
-    Returns:
-        pd.DataFrame: Aggregated cross-validation results with mean and std
-                     for each metric across folds.
-    """
-    print(f"[INFO] Starting cross-validation with {cv_folds} folds...")
-    cv_results = pd.DataFrame(
-        cross_validate(
-            model,
-            X_train,
-            y_train,
-            scoring=scoring,
-            cv=cv_folds,
-            return_train_score=True,
-            n_jobs=1,
-        )
-    ).agg(["mean", "std"]).round(3).T
-
-    print(f"[INFO] Cross-validation completed successfully")
-    return cv_results
-
-
-def save_cv_results(cv_results: pd.DataFrame, model_type: str, cv_output_folder: str) -> None:
-    """
-    Save cross-validation results to CSV file.
-
-    Args:
-        cv_results (pd.DataFrame): Aggregated cross-validation results.
-        model_type (str): Name of the model for filename.
-        cv_output_folder (str): Folder to save the CSV.
-
-    Returns:
-        None
-    """
-    cv_results_path = os.path.join(cv_output_folder, f"{model_type}_cv_results.csv")
-    cv_results.to_csv(cv_results_path)
-    print(f"[INFO] Saved CV results to {cv_results_path}")
-
-
-def perform_hyperparameter_optimization(
-    model,
-    X_train: pd.DataFrame,
-    y_train: pd.Series,
-    search_space: dict,
-    n_iter: int = HYPERPARAMETER_ITERATIONS
-):
-    """
-    Perform hyperparameter optimization using randomized search.
-
-    Searches over the parameter space to find the best hyperparameters
-    based on F1 score for the positive class.
-
-    Args:
-        model: Scikit-learn estimator to optimize.
-        X_train (pd.DataFrame): Training features.
-        y_train (pd.Series): Training labels.
-        search_space (dict): Parameter distributions to search.
-        n_iter (int): Number of parameter sets to evaluate (default: 30).
-
-    Returns:
-        RandomizedSearchCV: Fitted search object with best_estimator_ and best_params_.
-    """
-    print(f"[INFO] Starting hyperparameter optimization with {n_iter} iterations...")
-
-    search = RandomizedSearchCV(
-        estimator=model,
-        param_distributions=search_space,
-        n_iter=n_iter,
-        cv=CV_FOLDS,
-        scoring=make_scorer(
-            f1_score,
-            zero_division=0,
-            pos_label=TARGET_POSITIVE_LABEL
-        ),
-        n_jobs=1,
-        random_state=RANDOM_STATE,
-        return_train_score=True,
-        verbose=2,
-    )
-
-    print("[INFO] Fitting hyperparameter search...")
-    search.fit(X_train, y_train)
-
-    print(f"[INFO] Hyperparameter optimization completed successfully")
-    print(f"[INFO] Best params: {search.best_params_}")
-    print(f"[INFO] Best F1 score: {search.best_score_:.3f}")
-
-    return search
-
-
-def save_model(best_model, model_type: str, models_folder: str) -> None:
-    """
-    Save trained model to disk using joblib.
-
-    Args:
-        best_model: Fitted scikit-learn estimator.
-        model_type (str): Name of the model for filename.
-        models_folder (str): Folder to save the model pickle.
-
-    Returns:
-        None
-    """
-    model_path = os.path.join(models_folder, f"{model_type}_best.pkl")
-    joblib.dump(best_model, model_path)
-    print(f"[INFO] Saved best model to {model_path}")
 
 
 @click.command()
@@ -340,7 +97,7 @@ def main(model_type: str, train_path: str, cv_output_folder: str, models_folder:
 
     # Step 1: Load training data
     print(f"[STATUS] [{model_type.upper()}] Loading training data...")
-    X_train, y_train = load_training_data(train_path)
+    X_train, y_train = load_split_data(train_path, "Train")
 
     # Step 2: Define scoring metrics
     print(f"[STATUS] [{model_type.upper()}] Defining scoring metrics...")
