+++ scripts/08-test.py
@@ -19,174 +19,18 @@ import pandas as pd
 import joblib
 import matplotlib.pyplot as plt
 from sklearn.metrics import classification_report, ConfusionMatrixDisplay
+from src.load_model import load_model
+from src.load_split_data import load_split_data
+from src.generate_classification_report import generate_classification_report
+from src.prepare_report_dataframe import prepare_report_dataframe
+from src.save_classification_report import save_classification_report
+from src.generate_confusion_matrix import generate_confusion_matrix
 
 # Default configuration constants
 DEFAULT_MODEL_PATH = "../results/models/svc_best.pkl"
 DEFAULT_TEST_PATH = "../data/transformed/test_transformed.csv"
 DEFAULT_TABLES_FOLDER = "../results/tables"
 DEFAULT_CHARTS_FOLDER = "../results/charts"
-
-
-def load_model(model_path: str):
-    """
-    Load trained model from disk.
-
-    Args:
-        model_path (str): Path to the trained model pickle file.
-
-    Returns:
-        Trained scikit-learn estimator.
-
-    Raises:
-        FileNotFoundError: If the model file does not exist.
-    """
-    print(f"[INFO] Loading model")
-    best_model = joblib.load(model_path)
-    return best_model
-
-
-def load_test_data(test_path: str) -> tuple[pd.DataFrame, pd.Series]:
-    """
-    Load test data and extract features and labels.
-
-    Reads test data from CSV and validates that the required 'Class'
-    column is present.
-
-    Args:
-        test_path (str): Path to the test data CSV file.
-
-    Returns:
-        tuple[pd.DataFrame, pd.Series]: A tuple containing:
-            - X_test: Test features (all columns except 'Class')
-            - y_test: Test labels (the 'Class' column)
-
-    Raises:
-        ValueError: If 'Class' column is missing from the data.
-    """
-    print(f"[INFO] Loading test data")
-    test_df = pd.read_csv(test_path)
-
-    if "Class" not in test_df.columns:
-        raise ValueError("Test data must include a 'Class' column.")
-
-    X_test = test_df.drop(columns=["Class"])
-    y_test = test_df["Class"]
-
-    return X_test, y_test
-
-
-def generate_classification_report(model, X_test: pd.DataFrame, y_test: pd.Series) -> dict:
-    """
-    Generate classification report metrics.
-
-    Computes predictions and generates a detailed classification report
-    including precision, recall, and F1-score for each class.
-
-    Args:
-        model: Fitted scikit-learn estimator.
-        X_test (pd.DataFrame): Test features.
-        y_test (pd.Series): True test labels.
-
-    Returns:
-        dict: Classification report as a dictionary with metrics for each class.
-    """
-    print("[INFO] Generating classification report")
-    y_pred = model.predict(X_test)
-    report_dict = classification_report(y_test, y_pred, output_dict=True)
-    return report_dict
-
-
-def prepare_report_dataframe(report_dict: dict) -> pd.DataFrame:
-    """
-    Prepare classification report for saving to CSV.
-
-    Transforms the dictionary report into a structured DataFrame,
-    removing non-malicious class columns and organizing metrics.
-
-    Args:
-        report_dict (dict): Classification report dictionary.
-
-    Returns:
-        pd.DataFrame: Formatted report table ready for CSV export.
-    """
-    # Convert to DataFrame and remove averaged metrics
-    result_df = (
-        pd.DataFrame(report_dict)
-        .drop(["Benign", "macro avg", "weighted avg"], axis=1, errors="ignore")
-        .drop("support", errors="ignore")
-    )
-
-    # Keep only malicious class metrics and accuracy
-    result_table = result_df.copy().drop("accuracy", axis=1, errors="ignore")
-    result_table.columns = [""]
-
-    # Add accuracy row if present
-    if "accuracy" in result_df.index:
-        result_table.loc["test_accuracy"] = result_df["accuracy"].iloc[0]
-
-    return result_table
-
-
-def save_classification_report(
-    result_table: pd.DataFrame,
-    model_path: str,
-    tables_folder: str
-) -> None:
-    """
-    Save classification report to CSV file.
-
-    Args:
-        result_table (pd.DataFrame): Formatted classification report.
-        model_path (str): Path to the model (used to extract model name).
-        tables_folder (str): Folder to save the CSV.
-
-    Returns:
-        None
-    """
-    model_stem = os.path.splitext(os.path.basename(model_path))[0]
-    report_path = os.path.join(tables_folder, f"{model_stem}_class_report.csv")
-    result_table.T.to_csv(report_path)
-    print(f"[INFO] Saved classification report to {report_path}")
-
-
-def generate_confusion_matrix(
-    model,
-    X_test: pd.DataFrame,
-    y_test: pd.Series,
-    model_path: str,
-    charts_folder: str
-) -> None:
-    """
-    Generate and save confusion matrix visualization.
-
-    Creates a confusion matrix plot showing true vs predicted labels
-    and saves it as a PNG file.
-
-    Args:
-        model: Fitted scikit-learn estimator.
-        X_test (pd.DataFrame): Test features.
-        y_test (pd.Series): True test labels.
-        model_path (str): Path to the model (used to extract model name).
-        charts_folder (str): Folder to save the PNG chart.
-
-    Returns:
-        None
-    """
-    print("[INFO] Saving confusion matrix chart")
-
-    # Create confusion matrix display
-    disp = ConfusionMatrixDisplay.from_estimator(model, X_test, y_test)
-
-    # Extract model name for title and filename
-    model_stem = os.path.splitext(os.path.basename(model_path))[0]
-    plt.title(f"Confusion Matrix - {model_stem}")
-
-    # Save chart
-    chart_path = os.path.join(charts_folder, f"{model_stem}_confusion_matrix.png")
-    plt.savefig(chart_path, bbox_inches="tight")
-    plt.close()
-
-    print(f"[INFO] Saved confusion matrix chart to {chart_path}")
 
 
 @click.command()
@@ -243,7 +87,7 @@ def main(model_path: str, test_path: str, tables_folder: str, charts_folder: str
     best_model = load_model(model_path)
 
     # Step 2: Load test data
-    X_test, y_test = load_test_data(test_path)
+    X_test, y_test = load_split_data(test_path, "Test")
 
     # Step 3: Generate classification report
     report_dict = generate_classification_report(best_model, X_test, y_test)
