+++ src/analyze_class_balance.py
@@ -1,0 +1,31 @@
+import os
+import pandas as pd
+import matplotlib.pyplot as plt
+
+def analyze_class_balance(train: pd.DataFrame, tables_folder: str, charts_folder: str) -> None:
+    """
+    Analyze and visualize target class balance.
+
+    Generates both a CSV table with class proportions and a bar chart
+    visualization showing the distribution of the target variable.
+
+    Args:
+        train (pd.DataFrame): The training dataset (must contain 'Class' column).
+        tables_folder (str): Path to folder where CSV will be saved.
+        charts_folder (str): Path to folder where PNG chart will be saved.
+
+    Returns:
+        None
+    """
+    # Calculate class proportions
+    train_balance = train["Class"].value_counts(normalize=True).round(4)
+    balance_path_table = os.path.join(tables_folder, "train_class_balance.csv")
+    train_balance.to_csv(balance_path_table, header=["Proportion"])
+
+    # Generate and save bar chart
+    counts = train["Class"].value_counts()
+    counts.plot(kind="bar", title="Class Balance", rot=0)
+    balance_path_chart = os.path.join(charts_folder, "train_class_balance.png")
+    plt.tight_layout()
+    plt.savefig(balance_path_chart)
+    plt.close()
