+++ src/build_preprocessor.py
@@ -1,0 +1,48 @@
+from typing import List
+from sklearn.compose import ColumnTransformer
+from sklearn.pipeline import Pipeline
+from sklearn.preprocessing import StandardScaler, OneHotEncoder
+from sklearn.impute import SimpleImputer
+
+def build_preprocessor(numeric_cols: List[str], cat_cols: List[str]) -> ColumnTransformer:
+    """
+    Build a preprocessing pipeline for numeric and categorical features.
+
+    Creates a ColumnTransformer that applies different preprocessing steps
+    to numeric and categorical columns:
+    - Numeric: imputation (median) then standardization
+    - Categorical: imputation (mode) then one-hot encoding
+
+    Args:
+        numeric_cols (List[str]): List of numeric column names.
+        cat_cols (List[str]): List of categorical column names.
+
+    Returns:
+        ColumnTransformer: Sklearn preprocessor with separate pipelines
+                          for numeric and categorical features.
+    """
+    # Pipeline for numeric features: handle missing values then standardize
+    numeric_transformer = Pipeline(
+        steps=[
+            ("imputer", SimpleImputer(strategy="median")),  # Use median for robustness
+            ("scaler", StandardScaler()),  # Zero mean, unit variance
+        ]
+    )
+
+    # Pipeline for categorical features: handle missing values then encode
+    categorical_transformer = Pipeline(
+        steps=[
+            ("imputer", SimpleImputer(strategy="most_frequent")),  # Use mode for categories
+            ("onehot", OneHotEncoder(handle_unknown="ignore")),  # Convert to numeric
+        ]
+    )
+
+    # Combine both pipelines using ColumnTransformer
+    preprocessor = ColumnTransformer(
+        transformers=[
+            ("num", numeric_transformer, numeric_cols),
+            ("cat", categorical_transformer, cat_cols),
+        ]
+    )
+
+    return preprocessor
