+++ src/build_validation_schema.py
@@ -1,0 +1,352 @@
+import pandera.pandas as pa
+from pandera import Column, Check, DataFrameSchema
+
+# Validation parameter
+NULLNESS_THRESHOLD = 0.05  # Maximum proportion of null values allowed (5%)
+
+def build_validation_schema() -> pa.DataFrameSchema:
+    """
+    Build and return the data validation schema.
+
+    Defines the expected structure of the cleaned PDF malware dataset including
+    column names, data types, nullness constraints, and value ranges.
+
+    Returns:
+        pa.DataFrameSchema: Pandera schema object defining validation rules.
+    """
+    # Schema definition with column constraints and validations
+    schema = pa.DataFrameSchema(
+        {
+            "FileName": pa.Column(str, nullable=False),
+            "PdfSize": pa.Column(
+                float,
+                checks=[
+                    pa.Check.ge(0),
+                    pa.Check(
+                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                        element_wise=False,
+                        error=f"Too many null values in 'PdfSize' column.",
+                    ),
+                ],
+                nullable=True,
+                coerce=True,
+            ),
+            "MetadataSize": pa.Column(
+                float,
+                checks=[
+                    pa.Check.ge(0),
+                    pa.Check(
+                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                        element_wise=False,
+                        error=f"Too many null values in 'MetadataSize' column.",
+                    ),
+                ],
+                nullable=True,
+                coerce=True,
+            ),
+            "Pages": pa.Column(
+                float,
+                checks=[
+                    pa.Check.ge(0),
+                    pa.Check(
+                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                        element_wise=False,
+                        error=f"Too many null values in 'Pages' column.",
+                    ),
+                ],
+                nullable=True,
+                coerce=True,
+            ),
+            "XrefLength": pa.Column(
+                float,
+                checks=[
+                    pa.Check.ge(0),
+                    pa.Check(
+                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                        element_wise=False,
+                        error=f"Too many null values in 'XrefLength' column.",
+                    ),
+                ],
+                nullable=True,
+                coerce=True,
+            ),
+            "TitleCharacters": pa.Column(
+                float,
+                checks=[
+                    pa.Check.ge(0),
+                    pa.Check(
+                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                        element_wise=False,
+                        error=f"Too many null values in 'TitleCharacters' column.",
+                    ),
+                ],
+                nullable=True,
+                coerce=True,
+            ),
+            "isEncrypted": pa.Column(float, nullable=False),
+            "EmbeddedFiles": pa.Column(
+                float,
+                checks=[
+                    pa.Check.ge(0),
+                    pa.Check(
+                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                        element_wise=False,
+                        error=f"Too many null values in 'EmbeddedFiles' column.",
+                    ),
+                ],
+                nullable=True,
+                coerce=True,
+            ),
+            "Images": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'Images' column.",
+                ),
+                nullable=True,
+            ),
+            "Text": pa.Column(
+                str,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'Text' column.",
+                ),
+                nullable=True,
+            ),
+            "Header": pa.Column(
+                str,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'Header' column.",
+                ),
+                nullable=True,
+            ),
+            "Obj": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'Obj' column.",
+                ),
+                nullable=True,
+            ),
+            "Endobj": pa.Column(
+                float,
+                checks=[
+                    pa.Check.ge(0),
+                    pa.Check(
+                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                        element_wise=False,
+                        error=f"Too many null values in 'Endobj' column.",
+                    ),
+                ],
+                nullable=True,
+                coerce=True,
+            ),
+            "Stream": pa.Column(
+                float,
+                checks=[
+                    pa.Check.ge(0),
+                    pa.Check(
+                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                        element_wise=False,
+                        error=f"Too many null values in 'Stream' column.",
+                    ),
+                ],
+                nullable=True,
+                coerce=True,
+            ),
+            "Endstream": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'Endstream' column.",
+                ),
+                nullable=True,
+            ),
+            "Xref": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'Xref' column.",
+                ),
+                nullable=True,
+            ),
+            "Trailer": pa.Column(
+                float,
+                checks=[
+                    pa.Check.ge(0),
+                    pa.Check(
+                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                        element_wise=False,
+                        error=f"Too many null values in 'Trailer' column.",
+                    ),
+                ],
+                nullable=True,
+                coerce=True,
+            ),
+            "StartXref": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'StartXref' column.",
+                ),
+                nullable=True,
+            ),
+            "PageNo": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'PageNo' column.",
+                ),
+                nullable=True,
+            ),
+            "Encrypt": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'Encrypt' column.",
+                ),
+                nullable=True,
+                coerce=True,
+            ),
+            "ObjStm": pa.Column(
+                float,
+                checks=[
+                    pa.Check.ge(0),
+                    pa.Check(
+                        lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                        element_wise=False,
+                        error=f"Too many null values in 'ObjStm' column.",
+                    ),
+                ],
+                nullable=True,
+                coerce=True,
+            ),
+            "JS": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'JS' column.",
+                ),
+                nullable=True,
+            ),
+            "Javascript": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'Javascript' column.",
+                ),
+                nullable=True,
+            ),
+            "AA": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'AA' column.",
+                ),
+                nullable=True,
+            ),
+            "OpenAction": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'OpenAction' column.",
+                ),
+                nullable=True,
+            ),
+            "Acroform": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'Acroform' column.",
+                ),
+                nullable=True,
+            ),
+            "JBIG2Decode": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'JBIG2Decode' column.",
+                ),
+                nullable=True,
+            ),
+            "RichMedia": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'RichMedia' column.",
+                ),
+                nullable=True,
+            ),
+            "Launch": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'Launch' column.",
+                ),
+                nullable=True,
+            ),
+            "EmbeddedFile": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'EmbeddedFile' column.",
+                ),
+                nullable=True,
+            ),
+            "XFA": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'XFA' column.",
+                ),
+                nullable=True,
+            ),
+            "Colors": pa.Column(
+                float,
+                pa.Check(
+                    lambda s: s.isna().mean() <= NULLNESS_THRESHOLD,
+                    element_wise=False,
+                    error=f"Too many null values in 'Colors' column.",
+                ),
+                nullable=True,
+                coerce=True,
+            ),
+            "Class": pa.Column(
+                str,
+                pa.Check.isin(["Malicious", "Benign"]),
+                nullable=False,
+            ),
+        },
+        checks=[
+            pa.Check(
+                lambda clean_data: ~clean_data.duplicated().any(),
+                error="Duplicate rows found.",
+            ),
+            pa.Check(
+                lambda clean_data: ~(clean_data.isna().all(axis=1)).any(),
+                error="Empty rows found.",
+            ),
+        ],
+    )
+    return schema
+
