+++ src/check_outliers.py
@@ -1,0 +1,35 @@
+import pandas as pd
+import numpy as np
+from scipy import stats
+
+
+OUTLIER_Z_SCORE = 3  # Z-score threshold for outlier detection
+
+def check_outliers(clean_data: pd.DataFrame) -> bool:
+    """
+    Detect outliers using Z-score analysis on numeric columns.
+
+    Identifies anomalous values by computing Z-scores and flagging values
+    exceeding the threshold. Warns user if outliers are detected.
+
+    Args:
+        clean_data (pd.DataFrame): The DataFrame to check for outliers.
+
+    Returns:
+        bool: True if outliers are detected, False otherwise.
+    """
+    # Calculate Z-scores for numeric columns, excluding NaN values
+    numeric_data = clean_data.dropna().select_dtypes(include=np.number)
+
+    if numeric_data.empty:
+        print("No numeric data available for outlier detection")
+        return False
+
+    z_scores = np.abs(stats.zscore(numeric_data))
+    outliers = (z_scores > OUTLIER_Z_SCORE)
+
+    if outliers.any():
+        print(f"Warning: outliers detected (Z-score > {OUTLIER_Z_SCORE})")
+        return True
+
+    return False
