+++ src/clean_encrypted_column.py
@@ -1,0 +1,22 @@
+import pandas as pd
+import numpy as np
+
+def clean_encrypted_column(df: pd.DataFrame) -> None:
+    """
+    Clean the isEncrypted column: convert -1 to NaN, then binarize.
+
+    Replaces error code (-1) with NaN, then converts all values to binary
+    (0 or 1) where values > 0 are True. Modifies the DataFrame in-place.
+
+    Args:
+        df (pd.DataFrame): The DataFrame to modify. Must contain 'isEncrypted' column.
+
+    Returns:
+        None
+    """
+    if "isEncrypted" in df.columns:
+        # Replace error code with NaN
+        df.loc[df["isEncrypted"] == -1, "isEncrypted"] = np.nan
+
+        # Binarize: values > 0 become 1, others become 0
+        df["isEncrypted"] = (df["isEncrypted"] > 0).astype("float")
