+++ src/clean_text_header_columns.py
@@ -1,0 +1,29 @@
+import pandas as pd
+import numpy as np
+
+# Text/header columns requiring special error code handling
+TEXT_HEADER_COLS = ["Text", "Header"]
+
+def clean_text_header_columns(df: pd.DataFrame) -> None:
+    """
+    Clean Text and Header columns by replacing error codes with NaN.
+
+    Converts columns to string type, marks error codes (0, -1) as NaN,
+    then converts to categorical dtype. Modifies the DataFrame in-place.
+
+    Args:
+        df (pd.DataFrame): The DataFrame to modify.
+
+    Returns:
+        None
+    """
+    for col in TEXT_HEADER_COLS:
+        if col in df.columns:
+            # Convert to string type
+            df[col] = df[col].astype("string")
+
+            # Replace error sentinel values with NaN
+            df.loc[df[col].isin(["0", "-1"]), col] = np.nan
+
+            # Convert to categorical for memory efficiency
+            df[col] = df[col].astype("category")
