+++ src/compute_dtype_distribution.py
@@ -1,0 +1,23 @@
+import os
+import pandas as pd
+
+def compute_dtype_distribution(train: pd.DataFrame, tables_folder: str) -> None:
+    """
+    Compute and save data type distribution.
+
+    Analyzes the frequency of each data type (float, object, category, etc.)
+    in the dataset and saves to a CSV file.
+
+    Args:
+        train (pd.DataFrame): The training dataset.
+        tables_folder (str): Path to the folder where CSV will be saved.
+
+    Returns:
+        None
+    """
+    # Count occurrences of each data type
+    dtype_counts = train.dtypes.value_counts().reset_index()
+    dtype_counts.columns = ["Data Type", "Count"]
+
+    dtype_path = os.path.join(tables_folder, "dtype_counts.csv")
+    dtype_counts.to_csv(dtype_path, index=False)
