+++ src/compute_shape_stats.py
@@ -1,0 +1,23 @@
+import os
+import pandas as pd
+
+def compute_shape_stats(train: pd.DataFrame, tables_folder: str) -> None:
+    """
+    Compute and save dataset shape statistics.
+
+    Calculates the number of rows and columns in the training dataset
+    and saves to a CSV file.
+
+    Args:
+        train (pd.DataFrame): The training dataset.
+        tables_folder (str): Path to the folder where CSV will be saved.
+
+    Returns:
+        None
+    """
+    shape_df = pd.DataFrame({
+        "rows": [train.shape[0]],
+        "columns": [train.shape[1]]
+    })
+    shape_path = os.path.join(tables_folder, "train_shape.csv")
+    shape_df.to_csv(shape_path, index=False)
