+++ src/convert_categorical_to_numeric.py
@@ -1,0 +1,46 @@
+import pandas as pd
+
+# Columns that require categorical-to-numeric conversion
+CATEGORICAL_NUMERIC_COLS = [
+    "Obj",
+    "Endobj",
+    "Xref",
+    "StartXref",
+    "PageNo",
+    "Images",
+    "JS",
+    "Javascript",
+    "AA",
+    "Endstream",
+    "OpenAction",
+    "Acroform",
+    "JBIG2Decode",
+    "RichMedia",
+    "Launch",
+    "EmbeddedFile",
+    "XFA",
+]
+
+def convert_categorical_to_numeric(df: pd.DataFrame) -> None:
+    """
+    Convert categorical columns containing numeric strings to float type.
+
+    Extracts numeric values from categorical columns using regex, handling
+    negative numbers. Modifies the DataFrame in-place.
+
+    Args:
+        df (pd.DataFrame): The DataFrame to modify. Contains categorical columns
+                          with mixed string/numeric values.
+
+    Returns:
+        None
+    """
+    for col in CATEGORICAL_NUMERIC_COLS:
+        if col in df.columns:
+            # Convert to string, extract numeric pattern, convert to float
+            df[col] = (
+                df[col]
+                .astype(str)
+                .str.extract(r"(-?\d+)", expand=False)  # Extract numbers including negative
+                .astype("float")
+            )
