+++ src/create_directories.py
@@ -1,0 +1,18 @@
+import os
+
+def create_directories(data_save_folder: str, tables_folder: str) -> None:
+    """
+    Create necessary directories for data storage and results.
+
+    Creates directories if they don't already exist. Uses exist_ok=True to
+    prevent errors if directories already exist.
+
+    Args:
+        data_save_folder (str): Path to the folder where raw data will be stored.
+        tables_folder (str): Path to the folder where CSV results will be stored.
+
+    Returns:
+        None
+    """
+    os.makedirs(data_save_folder, exist_ok=True)
+    os.makedirs(tables_folder, exist_ok=True)
