+++ src/create_output_dataframes.py
@@ -1,0 +1,48 @@
+import pandas as pd
+from sklearn.compose import ColumnTransformer
+
+# Default target column name
+TARGET_COLUMN = "Class"
+
+def create_output_dataframes(
+    X_train_t: pd.DataFrame | None,
+    X_test_t: pd.DataFrame | None,
+    preprocessor: ColumnTransformer,
+    y_train: pd.Series | None,
+    y_test: pd.Series | None,
+    target_col: str = TARGET_COLUMN
+) -> tuple[pd.DataFrame, pd.DataFrame]:
+    """
+    Create output DataFrames from transformed features and original labels.
+
+    Converts transformed numpy arrays back to DataFrames with proper column names
+    and optionally appends the target variable.
+
+    Args:
+        X_train_t: Transformed training features (numpy array or sparse matrix).
+        X_test_t: Transformed test features (numpy array or sparse matrix).
+        preprocessor (ColumnTransformer): Fitted preprocessor with feature names.
+        y_train (pd.Series | None): Original training labels.
+        y_test (pd.Series | None): Original test labels.
+        target_col (str): Name of the target column (default: "Class").
+
+    Returns:
+        tuple[pd.DataFrame, pd.DataFrame]: A tuple containing:
+            - train_out: Training DataFrame with transformed features and labels
+            - test_out: Test DataFrame with transformed features and labels
+    """
+    # Get feature names from the fitted preprocessor
+    feature_names = preprocessor.get_feature_names_out()
+
+    # Convert transformed arrays to DataFrames
+    train_out = pd.DataFrame(X_train_t, columns=feature_names)
+    test_out = pd.DataFrame(X_test_t, columns=feature_names)
+
+    # Append target variable if present
+    if y_train is not None:
+        train_out[target_col] = y_train.reset_index(drop=True)
+
+    if y_test is not None:
+        test_out[target_col] = y_test.reset_index(drop=True)
+
+    return train_out, test_out
