+++ src/define_scoring_metrics.py
@@ -1,0 +1,34 @@
+from sklearn.metrics import make_scorer, precision_score, recall_score, f1_score
+
+# Model training parameters
+TARGET_POSITIVE_LABEL = "Malicious"  # Positive class for metrics
+
+def define_scoring_metrics() -> dict:
+    """
+    Define evaluation metrics for model cross-validation.
+
+    Creates a dictionary of scoring functions for evaluating model performance,
+    with emphasis on precision and recall for the "Malicious" class.
+
+    Returns:
+        dict: Dictionary mapping metric names to scoring functions or scorer objects.
+    """
+    scoring = {
+        "accuracy": "accuracy",  # Proportion of correct predictions
+        "precision": make_scorer(
+            precision_score,
+            zero_division=0,
+            pos_label=TARGET_POSITIVE_LABEL
+        ),  # True positives / all positive predictions
+        "recall": make_scorer(
+            recall_score,
+            zero_division=0,
+            pos_label=TARGET_POSITIVE_LABEL
+        ),  # True positives / all actual positives
+        "f1": make_scorer(
+            f1_score,
+            zero_division=0,
+            pos_label=TARGET_POSITIVE_LABEL
+        ),  # Harmonic mean of precision and recall
+    }
+    return scoring
