+++ src/download_and_copy_dataset.py
@@ -1,0 +1,47 @@
+import os
+import glob
+import shutil
+import pandas as pd
+import kagglehub
+
+def download_and_copy_dataset(link: str, data_save_folder: str) -> tuple[pd.DataFrame, str]:
+    """
+    Download dataset from Kaggle and copy to local storage.
+
+    Downloads the dataset using kagglehub and copies the first parquet file found
+    to the specified data folder. Validates that the copied file is in parquet format.
+
+    Args:
+        link (str): Kaggle dataset link in format "username/dataset-name".
+        data_save_folder (str): Path to the folder where the parquet file will be saved.
+
+    Returns:
+        tuple[pd.DataFrame, str]: A tuple containing:
+            - The loaded DataFrame from the parquet file
+            - The path to the saved parquet file
+
+    Raises:
+        AssertionError: If the saved file is not in parquet format.
+        IndexError: If no parquet files are found in the downloaded dataset.
+    """
+    # Download dataset from Kaggle - returns the local path to downloaded files
+    downloaded_path = kagglehub.dataset_download(link)
+
+    # Find the first parquet file in the downloaded dataset
+    downloaded_parquet_files = glob.glob(os.path.join(downloaded_path, "*.parquet"))
+    if not downloaded_parquet_files:
+        raise FileNotFoundError(f"No parquet files found in {downloaded_path}")
+
+    downloaded_parquet = downloaded_parquet_files[0]
+
+    # Copy the parquet file to our designated data folder
+    data_path = os.path.join(data_save_folder, "PDFMalware2022.parquet")
+    shutil.copy(downloaded_parquet, data_path)
+
+    # Validate that the saved file is indeed in parquet format
+    assert os.path.splitext(data_path)[1] == ".parquet", \
+        f"Expected parquet file, but got {os.path.splitext(data_path)[1]}"
+
+    # Load and return the dataset
+    data = pd.read_parquet(data_path, engine="pyarrow")
+    return data, data_path
