+++ src/export_data_summary.py
@@ -1,0 +1,37 @@
+import os
+import pandas as pd
+
+def export_data_summary(data: pd.DataFrame, tables_folder: str, prefix: str = "data") -> None:
+    """
+    Export data information and statistical summaries to CSV files.
+
+    Generates two CSV files:
+        1. {prefix}_info.csv: Column data types and null/non-null counts
+        2. {prefix}_describe.csv: Statistical summaries (mean, std, min, max, etc.)
+
+    Args:
+        data (pd.DataFrame): The DataFrame to analyze and export statistics for.
+        tables_folder (str): Path to the folder where CSV files will be saved.
+        prefix (str): Prefix for the name of output files (default: "data").
+
+    Returns:
+        None
+    """
+    # ensure tables folder exists so it doesn't break if it doesn't exist
+    os.makedirs(tables_folder, exist_ok=True)
+
+    # Export column-level information (data types and null counts)
+    info_df = pd.DataFrame(
+        {
+            "dtype": data.dtypes,
+            "non_null_count": data.notna().sum(),
+            "null_count": data.isna().sum(),
+        }
+    )
+    info_path = os.path.join(tables_folder, f"{prefix}_info.csv")
+    info_df.to_csv(info_path)
+
+    # Export descriptive statistics (mean, std, quartiles, etc.)
+    descr_df = data.describe()
+    descr_path = os.path.join(tables_folder, f"{prefix}_describe.csv")
+    descr_df.to_csv(descr_path)
