+++ src/extract_features.py
@@ -1,0 +1,23 @@
+import pandas as pd
+
+def extract_features(data: pd.DataFrame) -> pd.DataFrame:
+    """
+    Extract features by removing non-predictive identifier columns.
+
+    Removes columns like 'FileName' that are identifiers rather than features.
+    If FileName is not present, returns the data unchanged.
+
+    Args:
+        data (pd.DataFrame): The full dataset with all columns.
+
+    Returns:
+        pd.DataFrame: DataFrame with identifier columns removed.
+    """
+    if "FileName" in data.columns:
+        # Remove identifier column; keep all features and target
+        features = data.drop("FileName", axis=1)
+    else:
+        # No identifier columns to remove
+        features = data
+
+    return features
