+++ src/extract_features_and_labels.py
@@ -1,0 +1,37 @@
+import pandas as pd
+
+# Default target column name
+TARGET_COLUMN = "Class"
+
+def extract_features_and_labels(
+    train_df: pd.DataFrame,
+    test_df: pd.DataFrame,
+    target_col: str = TARGET_COLUMN
+) -> tuple[pd.DataFrame, pd.Series | None, pd.DataFrame, pd.Series | None]:
+    """
+    Extract features and labels from train/test DataFrames.
+
+    Separates the target variable from features. If the target column is absent,
+    returns None for the labels.
+
+    Args:
+        train_df (pd.DataFrame): The training DataFrame.
+        test_df (pd.DataFrame): The test DataFrame.
+        target_col (str): Name of the target column (default: "Class").
+
+    Returns:
+        tuple: A tuple containing:
+            - X_train: Training features
+            - y_train: Training labels (or None if not present)
+            - X_test: Test features
+            - y_test: Test labels (or None if not present)
+    """
+    # Extract features and labels from training data
+    X_train = train_df.drop(columns=[target_col], errors="ignore")
+    y_train = train_df[target_col] if target_col in train_df.columns else None
+
+    # Extract features and labels from test data
+    X_test = test_df.drop(columns=[target_col], errors="ignore")
+    y_test = test_df[target_col] if target_col in test_df.columns else None
+
+    return X_train, y_train, X_test, y_test
