+++ src/fit_and_transform_data.py
@@ -1,0 +1,32 @@
+import pandas as pd
+from sklearn.compose import ColumnTransformer
+
+def fit_and_transform_data(
+    preprocessor: ColumnTransformer,
+    X_train: pd.DataFrame,
+    X_test: pd.DataFrame
+) -> tuple[pd.DataFrame, pd.DataFrame]:
+    """
+    Fit preprocessor on training data and transform both train and test sets.
+
+    Applies the train-fit pattern: learns statistics from training data
+    and applies the same transformation to both training and test sets.
+
+    Args:
+        preprocessor (ColumnTransformer): The preprocessing pipeline.
+        X_train (pd.DataFrame): Training features.
+        X_test (pd.DataFrame): Test features.
+
+    Returns:
+        tuple[pd.DataFrame, pd.DataFrame]: A tuple containing:
+            - X_train_transformed: Preprocessed training features
+            - X_test_transformed: Preprocessed test features
+    """
+    # Fit preprocessor on training data
+    preprocessor.fit(X_train)
+
+    # Transform both training and test sets
+    X_train_t = preprocessor.transform(X_train)
+    X_test_t = preprocessor.transform(X_test)
+
+    return X_train_t, X_test_t
