+++ src/generate_classification_report.py
@@ -1,0 +1,22 @@
+from sklearn.metrics import classification_report
+import pandas as pd 
+
+def generate_classification_report(model, X_test: pd.DataFrame, y_test: pd.Series) -> dict:
+    """
+    Generate classification report metrics.
+
+    Computes predictions and generates a detailed classification report
+    including precision, recall, and F1-score for each class.
+
+    Args:
+        model: Fitted scikit-learn estimator.
+        X_test (pd.DataFrame): Test features.
+        y_test (pd.Series): True test labels.
+
+    Returns:
+        dict: Classification report as a dictionary with metrics for each class.
+    """
+    print("[INFO] Generating classification report")
+    y_pred = model.predict(X_test)
+    report_dict = classification_report(y_test, y_pred, output_dict=True)
+    return report_dict
