+++ src/generate_confusion_matrix.py
@@ -1,0 +1,44 @@
+import os
+import pandas as pd
+import matplotlib.pyplot as plt
+from sklearn.metrics import ConfusionMatrixDisplay
+
+
+def generate_confusion_matrix(
+    model,
+    X_test: pd.DataFrame,
+    y_test: pd.Series,
+    model_path: str,
+    charts_folder: str
+) -> None:
+    """
+    Generate and save confusion matrix visualization.
+
+    Creates a confusion matrix plot showing true vs predicted labels
+    and saves it as a PNG file.
+
+    Args:
+        model: Fitted scikit-learn estimator.
+        X_test (pd.DataFrame): Test features.
+        y_test (pd.Series): True test labels.
+        model_path (str): Path to the model (used to extract model name).
+        charts_folder (str): Folder to save the PNG chart.
+
+    Returns:
+        None
+    """
+    print("[INFO] Saving confusion matrix chart")
+
+    # Create confusion matrix display
+    disp = ConfusionMatrixDisplay.from_estimator(model, X_test, y_test)
+
+    # Extract model name for title and filename
+    model_stem = os.path.splitext(os.path.basename(model_path))[0]
+    plt.title(f"Confusion Matrix - {model_stem}")
+
+    # Save chart
+    chart_path = os.path.join(charts_folder, f"{model_stem}_confusion_matrix.png")
+    plt.savefig(chart_path, bbox_inches="tight")
+    plt.close()
+
+    print(f"[INFO] Saved confusion matrix chart to {chart_path}")
