+++ src/generate_correlation_chart.py
@@ -1,0 +1,21 @@
+import os
+import pandas as pd
+import altair_ally as aly
+
+def generate_correlation_chart(train: pd.DataFrame, charts_folder: str) -> None:
+    """
+    Generate correlation matrix visualization.
+
+    Creates an interactive HTML heatmap showing pairwise correlations
+    between all numeric features.
+
+    Args:
+        train (pd.DataFrame): The training dataset with numeric features.
+        charts_folder (str): Path to folder where HTML chart will be saved.
+
+    Returns:
+        None
+    """
+    corr_chart = aly.corr(train)
+    corr_path = os.path.join(charts_folder, "train_correlations.html")
+    corr_chart.save(corr_path)
