+++ src/generate_descriptive_stats.py
@@ -1,0 +1,19 @@
+import os
+import pandas as pd
+
+def generate_descriptive_stats(train: pd.DataFrame, tables_folder: str) -> None:
+    """
+    Generate and save descriptive statistics for all features.
+
+    Computes statistical summaries (mean, std, min, max, quartiles, etc.)
+    for all columns in the dataset.
+
+    Args:
+        train (pd.DataFrame): The training dataset.
+        tables_folder (str): Path to folder where CSV will be saved.
+
+    Returns:
+        None
+    """
+    train_descr_path = os.path.join(tables_folder, "train_describe.csv")
+    train.describe(include="all").round(2).to_csv(train_descr_path)
