+++ src/generate_distribution_charts.py
@@ -1,0 +1,26 @@
+import os
+import pandas as pd
+import altair as alt
+import altair_ally as aly
+
+def generate_distribution_charts(train: pd.DataFrame, charts_folder: str) -> None:
+    """
+    Generate univariate distribution charts for all features.
+
+    Creates interactive HTML charts showing the distribution of each feature,
+    colored by the target class.
+
+    Args:
+        train (pd.DataFrame): The training dataset (must contain 'Class' column).
+        charts_folder (str): Path to folder where HTML chart will be saved.
+
+    Returns:
+        None
+    """
+    # Disable row limit for Altair to handle large datasets
+    alt.data_transformers.disable_max_rows()
+
+    # Generate univariate distribution chart with 4 columns layout
+    univar_chart = aly.dist(train, color='Class', columns=4)
+    univar_path = os.path.join(charts_folder, "train_univar_distribution.html")
+    univar_chart.save(univar_path)
