+++ src/get_model_and_search_space.py
@@ -1,0 +1,62 @@
+from sklearn.dummy import DummyClassifier
+from sklearn.linear_model import LogisticRegression
+from sklearn.svm import SVC
+from scipy.stats import loguniform
+
+
+# Model training parameter
+RANDOM_STATE = 42  # Fixed seed for reproducibility
+
+def get_model_and_search_space(model_type: str):
+    """
+    Get model instance and hyperparameter search space for a given model type.
+
+    Returns a scikit-learn estimator and corresponding hyperparameter search space
+    for use with RandomizedSearchCV. Each model has different hyperparameters
+    and search ranges.
+
+    Args:
+        model_type (str): Type of model to instantiate. Options:
+            - "dummy": Dummy classifier baseline
+            - "logisticreg": Logistic regression
+            - "svc": Support vector classifier
+
+    Returns:
+        tuple: A tuple containing:
+            - model: Instantiated scikit-learn estimator
+            - search_space (dict): Hyperparameter distributions for RandomizedSearchCV
+
+    Raises:
+        ValueError: If model_type is not supported.
+    """
+    if model_type == "dummy":
+        # Baseline classifier that makes simple predictions
+        model = DummyClassifier(random_state=RANDOM_STATE)
+        search_space = {
+            "strategy": ["most_frequent", "prior"]  # Baseline strategies
+        }
+
+    elif model_type == "logisticreg":
+        # Logistic regression classifier
+        model = LogisticRegression(max_iter=1000, random_state=RANDOM_STATE)
+        search_space = {
+            "C": loguniform(1e-3, 1e3),  # Inverse regularization strength
+            "class_weight": [None, "balanced"],  # Handle class imbalance
+        }
+
+    elif model_type == "svc":
+        # Support Vector Classifier
+        model = SVC(random_state=RANDOM_STATE)
+        search_space = {
+            "C": loguniform(1e-3, 1e3),  # Regularization parameter
+            "gamma": loguniform(1e-4, 1e1),  # Kernel coefficient
+            "class_weight": [None, "balanced"],  # Handle class imbalance
+        }
+
+    else:
+        raise ValueError(
+            f"Unsupported model type: {model_type}. "
+            f"Supported types: 'dummy', 'logisticreg', 'svc'"
+        )
+
+    return model, search_space
