+++ src/handle_categorical_nulls.py
@@ -1,0 +1,23 @@
+import pandas as pd
+import numpy as np
+
+def handle_categorical_nulls(df: pd.DataFrame) -> None:
+    """
+    Normalize null handling in categorical columns.
+
+    Converts categorical columns to object type and ensures NaN values are
+    consistently represented. Modifies the DataFrame in-place.
+
+    Args:
+        df (pd.DataFrame): The DataFrame to modify. Contains categorical columns.
+
+    Returns:
+        None
+    """
+    # Identify categorical columns
+    cat_cols = [col for col in df.columns if df[col].dtype == "category"]
+
+    # Convert to object and normalize NaN representation
+    for col in cat_cols:
+        df[col] = df[col].astype(object)
+        df.loc[df[col].isna(), col] = np.nan
