+++ src/handle_numeric_error_codes.py
@@ -1,0 +1,28 @@
+import pandas as pd
+import numpy as np
+
+# Text/header columns requiring special error code handling
+TEXT_HEADER_COLS = ["Text", "Header"]
+
+def handle_numeric_error_codes(df: pd.DataFrame) -> None:
+    """
+    Replace error codes with NaN in all numeric columns.
+
+    Iterates through numeric columns and replaces -1 (common error sentinel)
+    with NaN. Modifies the DataFrame in-place.
+
+    Args:
+        df (pd.DataFrame): The DataFrame to modify. Contains numeric columns with -1 errors.
+
+    Returns:
+        None
+    """
+    # Identify all numeric columns
+    numeric_cols = [
+        col for col in df.columns
+        if df[col].dtype != "category" and col not in TEXT_HEADER_COLS
+    ]
+
+    # Replace error code -1 with NaN
+    for col in numeric_cols:
+        df.loc[df[col] == -1, col] = np.nan
