+++ src/infer_feature_groups.py
@@ -1,0 +1,32 @@
+import pandas as pd
+
+# Default target column name
+TARGET_COLUMN = "Class"
+
+def infer_feature_groups(df: pd.DataFrame, target_col: str = TARGET_COLUMN) -> tuple[list[str], list[str]]:
+    """
+    Automatically infer numeric and categorical feature columns.
+
+    Separates features into numeric and categorical types, excluding the
+    target column. Uses pandas dtype inference to classify columns.
+
+    Args:
+        df (pd.DataFrame): The DataFrame to analyze for feature types.
+        target_col (str): Name of the target column to exclude (default: "Class").
+
+    Returns:
+        tuple[list[str], list[str]]: A tuple containing:
+            - numeric column names
+            - categorical column names (object or category dtype)
+    """
+    # Remove target column for feature analysis
+    feature_df = df.drop(columns=[target_col], errors="ignore")
+
+    # Identify categorical columns (object or category dtype)
+    cat_cols = feature_df.select_dtypes(include=["object", "category"]).columns.tolist()
+
+    # Numeric columns are everything else
+    num_cols = feature_df.columns.difference(cat_cols).tolist()
+
+    return num_cols, cat_cols
+
