+++ src/load_data.py
@@ -1,0 +1,18 @@
+import os
+import pandas as pd
+
+def load_data(data_path: str) -> pd.DataFrame:
+    """
+    Load data from CSV file.
+
+    Args:
+        data_path (str): Path to the cleaned data CSV file.
+
+    Returns:
+        pd.DataFrame: The loaded DataFrame.
+
+    Raises:
+        FileNotFoundError: If the specified file does not exist.
+    """
+    data = pd.read_csv(data_path)
+    return data
