+++ src/load_model.py
@@ -1,0 +1,19 @@
+import os
+import joblib
+
+def load_model(model_path: str):
+    """
+    Load trained model from disk.
+
+    Args:
+        model_path (str): Path to the trained model pickle file.
+
+    Returns:
+        Trained scikit-learn estimator.
+
+    Raises:
+        FileNotFoundError: If the model file does not exist.
+    """
+    print(f"[INFO] Loading model")
+    best_model = joblib.load(model_path)
+    return best_model
