+++ src/load_raw_data.py
@@ -1,0 +1,20 @@
+import os
+import pandas as pd
+
+def load_raw_data(data_import_folder: str, import_file_name: str) -> pd.DataFrame:
+    """
+    Load raw parquet data from disk.
+
+    Args:
+        data_import_folder (str): Path to the folder containing the parquet file.
+        import_file_name (str): Name of the parquet file to load.
+
+    Returns:
+        pd.DataFrame: The raw dataset loaded from the parquet file.
+
+    Raises:
+        FileNotFoundError: If the specified parquet file does not exist.
+    """
+    data_path = os.path.join(data_import_folder, import_file_name)
+    raw_data = pd.read_parquet(data_path, engine="pyarrow")
+    return raw_data
