+++ src/load_split_data.py
@@ -1,0 +1,33 @@
+import pandas as pd
+
+def load_split_data(path: str, type: str) -> tuple[pd.DataFrame, pd.Series]:
+    """
+    Load training/test data and extract features and labels.
+
+    Reads training/test data from CSV and validates that the required 'Class'
+    column is present.
+
+    Args:
+        path (str): Path to the training CSV file.
+        type (str): type of data (Train or Test)
+
+    Returns:
+        tuple[pd.DataFrame, pd.Series]: A tuple containing:
+            - X_{type}: Training/Test features (all columns except 'Class')
+            - y_{type}: Training/Test labels (the 'Class' column)
+
+    Raises:
+        ValueError: If 'Class' column is missing from the data.
+    """
+    print(f"[INFO] Loading {type} data")
+    df = pd.read_csv(path)
+
+    if "Class" not in df.columns:
+        raise ValueError(f"{type} data must include a 'Class' column.")
+
+    X_train = df.drop(columns=["Class"])
+    y_train = df["Class"]
+
+    print(f"[INFO] Loaded {type} data. Rows: {len(df)}, Features: {X_train.shape[1]}")
+
+    return X_train, y_train
