+++ src/perform_cross_validation.py
@@ -1,0 +1,46 @@
+import pandas as pd
+from sklearn.model_selection import cross_validate, RandomizedSearchCV
+
+
+# Model training parameter
+CV_FOLDS = 5  # Number of cross-validation folds
+
+def perform_cross_validation(
+    model,
+    X_train: pd.DataFrame,
+    y_train: pd.Series,
+    scoring: dict,
+    cv_folds: int = CV_FOLDS
+) -> pd.DataFrame:
+    """
+    Perform cross-validation and return aggregated results.
+
+    Evaluates model performance using k-fold cross-validation, computing
+    both training and validation scores for each metric.
+
+    Args:
+        model: Scikit-learn estimator to evaluate.
+        X_train (pd.DataFrame): Training features.
+        y_train (pd.Series): Training labels.
+        scoring (dict): Dictionary of scoring metrics.
+        cv_folds (int): Number of cross-validation folds (default: 5).
+
+    Returns:
+        pd.DataFrame: Aggregated cross-validation results with mean and std
+                     for each metric across folds.
+    """
+    print(f"[INFO] Starting cross-validation with {cv_folds} folds...")
+    cv_results = pd.DataFrame(
+        cross_validate(
+            model,
+            X_train,
+            y_train,
+            scoring=scoring,
+            cv=cv_folds,
+            return_train_score=True,
+            n_jobs=1,
+        )
+    ).agg(["mean", "std"]).round(3).T
+
+    print(f"[INFO] Cross-validation completed successfully")
+    return cv_results
