+++ src/perform_hyperparameter_optimization.py
@@ -1,0 +1,59 @@
+import pandas as pd
+from sklearn.model_selection import RandomizedSearchCV
+from sklearn.metrics import make_scorer, f1_score
+
+# Model training parameters
+CV_FOLDS = 5  # Number of cross-validation folds
+HYPERPARAMETER_ITERATIONS = 5  # Number of hyperparameter combinations to try
+RANDOM_STATE = 42  # Fixed seed for reproducibility
+TARGET_POSITIVE_LABEL = "Malicious"  # Positive class for metrics
+
+def perform_hyperparameter_optimization(
+    model,
+    X_train: pd.DataFrame,
+    y_train: pd.Series,
+    search_space: dict,
+    n_iter: int = HYPERPARAMETER_ITERATIONS
+):
+    """
+    Perform hyperparameter optimization using randomized search.
+
+    Searches over the parameter space to find the best hyperparameters
+    based on F1 score for the positive class.
+
+    Args:
+        model: Scikit-learn estimator to optimize.
+        X_train (pd.DataFrame): Training features.
+        y_train (pd.Series): Training labels.
+        search_space (dict): Parameter distributions to search.
+        n_iter (int): Number of parameter sets to evaluate (default: 30).
+
+    Returns:
+        RandomizedSearchCV: Fitted search object with best_estimator_ and best_params_.
+    """
+    print(f"[INFO] Starting hyperparameter optimization with {n_iter} iterations...")
+
+    search = RandomizedSearchCV(
+        estimator=model,
+        param_distributions=search_space,
+        n_iter=n_iter,
+        cv=CV_FOLDS,
+        scoring=make_scorer(
+            f1_score,
+            zero_division=0,
+            pos_label=TARGET_POSITIVE_LABEL
+        ),
+        n_jobs=1,
+        random_state=RANDOM_STATE,
+        return_train_score=True,
+        verbose=2,
+    )
+
+    print("[INFO] Fitting hyperparameter search...")
+    search.fit(X_train, y_train)
+
+    print(f"[INFO] Hyperparameter optimization completed successfully")
+    print(f"[INFO] Best params: {search.best_params_}")
+    print(f"[INFO] Best F1 score: {search.best_score_:.3f}")
+
+    return search
