+++ src/prepare_report_dataframe.py
@@ -1,0 +1,31 @@
+import pandas as pd
+
+def prepare_report_dataframe(report_dict: dict) -> pd.DataFrame:
+    """
+    Prepare classification report for saving to CSV.
+
+    Transforms the dictionary report into a structured DataFrame,
+    removing non-malicious class columns and organizing metrics.
+
+    Args:
+        report_dict (dict): Classification report dictionary.
+
+    Returns:
+        pd.DataFrame: Formatted report table ready for CSV export.
+    """
+    # Convert to DataFrame and remove averaged metrics
+    result_df = (
+        pd.DataFrame(report_dict)
+        .drop(["Benign", "macro avg", "weighted avg"], axis=1, errors="ignore")
+        .drop("support", errors="ignore")
+    )
+
+    # Keep only malicious class metrics and accuracy
+    result_table = result_df.copy().drop("accuracy", axis=1, errors="ignore")
+    result_table.columns = [""]
+
+    # Add accuracy row if present
+    if "accuracy" in result_df.index:
+        result_table.loc["test_accuracy"] = result_df["accuracy"].iloc[0]
+
+    return result_table
