+++ src/save_classification_report.py
@@ -1,0 +1,23 @@
+import os
+import pandas as pd
+
+def save_classification_report(
+    result_table: pd.DataFrame,
+    model_path: str,
+    tables_folder: str
+) -> None:
+    """
+    Save classification report to CSV file.
+
+    Args:
+        result_table (pd.DataFrame): Formatted classification report.
+        model_path (str): Path to the model (used to extract model name).
+        tables_folder (str): Folder to save the CSV.
+
+    Returns:
+        None
+    """
+    model_stem = os.path.splitext(os.path.basename(model_path))[0]
+    report_path = os.path.join(tables_folder, f"{model_stem}_class_report.csv")
+    result_table.T.to_csv(report_path)
+    print(f"[INFO] Saved classification report to {report_path}")
