+++ src/save_cv_results.py
@@ -1,0 +1,18 @@
+import os
+import pandas as pd
+
+def save_cv_results(cv_results: pd.DataFrame, model_type: str, cv_output_folder: str) -> None:
+    """
+    Save cross-validation results to CSV file.
+
+    Args:
+        cv_results (pd.DataFrame): Aggregated cross-validation results.
+        model_type (str): Name of the model for filename.
+        cv_output_folder (str): Folder to save the CSV.
+
+    Returns:
+        None
+    """
+    cv_results_path = os.path.join(cv_output_folder, f"{model_type}_cv_results.csv")
+    cv_results.to_csv(cv_results_path)
+    print(f"[INFO] Saved CV results to {cv_results_path}")
