+++ src/save_data.py
@@ -1,0 +1,21 @@
+import os
+import pandas as pd
+
+def save_data(data: pd.DataFrame, data_export_folder: str, prefix: str) -> None:
+    """
+    Save data to CSV file.
+
+    Args:
+        data (pd.DataFrame): The DataFrame to save.
+        data_export_folder (str): Path to the folder where the CSV will be saved.
+        prefix (str): the name of the file that will be saved as (e.g. prefix.csv).
+
+    Returns:
+        None
+    """
+    # Create output directory if needed
+    os.makedirs(data_export_folder, exist_ok=True)
+
+    data_path = os.path.join(data_export_folder, f"{prefix}.csv")
+    
+    data.to_csv(data_path, index=False)
