+++ src/save_model.py
@@ -1,0 +1,18 @@
+import os
+import joblib
+
+def save_model(best_model, model_type: str, models_folder: str) -> None:
+    """
+    Save trained model to disk using joblib.
+
+    Args:
+        best_model: Fitted scikit-learn estimator.
+        model_type (str): Name of the model for filename.
+        models_folder (str): Folder to save the model pickle.
+
+    Returns:
+        None
+    """
+    model_path = os.path.join(models_folder, f"{model_type}_best.pkl")
+    joblib.dump(best_model, model_path)
+    print(f"[INFO] Saved best model to {model_path}")
