+++ src/split_data.py
@@ -1,0 +1,29 @@
+import pandas as pd
+from sklearn.model_selection import train_test_split
+
+# Dataset splitting parameters
+TEST_SPLIT_SIZE = 0.2  # 80/20 train/test split
+RANDOM_STATE = 42  # Fixed seed for reproducibility
+
+def split_data(features: pd.DataFrame) -> tuple[pd.DataFrame, pd.DataFrame]:
+    """
+    Perform stratified train/test split on the feature data.
+
+    Splits data with 80% for training and 20% for testing using a fixed
+    random state to ensure reproducibility.
+
+    Args:
+        features (pd.DataFrame): The feature DataFrame to split. Should include
+                                the target column 'Class' if present.
+
+    Returns:
+        tuple[pd.DataFrame, pd.DataFrame]: A tuple containing:
+            - train: Training data (80%)
+            - test: Test data (20%)
+    """
+    train, test = train_test_split(
+        features,
+        test_size=TEST_SPLIT_SIZE,
+        random_state=RANDOM_STATE
+    )
+    return train, test
