+++ src/validate_class_balance.py
@@ -1,0 +1,35 @@
+import pandas as pd
+import numpy as np
+
+# Validation parameter
+CLASS_BALANCE_THRESHOLD = 0.15  # Maximum class imbalance (15% difference)
+
+def validate_class_balance(clean_data: pd.DataFrame) -> bool:
+    """
+    Validate that target class distribution meets balance requirements.
+
+    Checks that the difference in proportions between the two classes
+    does not exceed the defined threshold (default 15%).
+
+    Args:
+        clean_data (pd.DataFrame): The DataFrame containing the 'Class' column.
+
+    Returns:
+        bool: True if balance is acceptable, False if threshold is exceeded.
+    """
+    # Get class distribution as normalized proportions
+    clean_data_balance = clean_data["Class"].value_counts(normalize=True)
+
+    # Calculate difference between class proportions
+    class_diff = np.abs(clean_data_balance.iloc[0] - clean_data_balance.iloc[1])
+
+    if class_diff >= CLASS_BALANCE_THRESHOLD:
+        print(
+            f"Target variable distribution check failed. "
+            f"The count difference between classes ({class_diff:.2%}) "
+            f"exceeds the threshold ({CLASS_BALANCE_THRESHOLD:.2%})."
+        )
+        return False
+
+    return True
+
