+++ src/validate_schema.py
@@ -1,0 +1,28 @@
+import pandas as pd
+import pandera.pandas as pa
+import logging
+import json
+
+def validate_schema(clean_data: pd.DataFrame, schema: pa.DataFrameSchema) -> bool:
+    """
+    Validate data against the defined schema.
+
+    Performs schema validation including column names, data types, constraints,
+    duplicates, and empty rows. Logs validation errors if any are found.
+
+    Args:
+        clean_data (pd.DataFrame): The DataFrame to validate.
+        schema (pa.DataFrameSchema): The validation schema to apply.
+
+    Returns:
+        bool: True if validation passes, False otherwise.
+    """
+    try:
+        schema.validate(clean_data, lazy=True)
+        print("All data validation checks are passed!")
+        return True
+    except pa.errors.SchemaErrors as e:
+        print("Data validation checks failed, please see above")
+        error_message = json.dumps(e.message, indent=2)
+        logging.error("\n" + error_message)
+        return False
